feat: refactor ime structure
