+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
 	<uses-permission android:name="android.permission.VIBRATE" />
 
 	<application
-		android:name=".TeogeulApplication"
+		android:name=".Application"
 		android:icon="@mipmap/ic_launcher"
 		android:label="@string/app_name"
 		android:allowBackup="true"
@@ -17,7 +17,7 @@
 		android:dataExtractionRules="@xml/data_extraction_rules">
 		<service
 			android:id="@+id/teogeul_korean"
-			android:name="io.github.lens0021.teogeul.TeogeulKOKR"
+			android:name="io.github.lens0021.teogeul.IMEService"
 			android:label="@string/ime_name"
 			android:permission="android.permission.BIND_INPUT_METHOD"
 			android:exported="true">
@@ -31,7 +31,7 @@
 		</service>
 
 		<activity
-			android:name="io.github.lens0021.teogeul.ui.TeogeulSettingsActivity"
+			android:name="io.github.lens0021.teogeul.ui.SettingsActivity"
 			android:label="@string/preference_ime_setting_app"
 			android:taskAffinity=""
 			android:exported="true">
@@ -42,7 +42,7 @@
 			</intent-filter>
 		</activity>
 		<activity
-			android:name="io.github.lens0021.teogeul.KOKR.ListLangKeyActionDialogActivity"
+			android:name="io.github.lens0021.teogeul.korean.LangKeyActionDialog"
 			android:theme="@android:style/Theme.Translucent.NoTitleBar"
 			android:launchMode="singleInstance"
 			android:excludeFromRecents="true"
