+++ app/src/main/java/io/github/lens0021/teogeul/Application.kt
@@ -1,7 +1,6 @@
 package io.github.lens0021.teogeul
 
-import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
 
 @HiltAndroidApp
-class TeogeulApplication : Application()
+class Application : android.app.Application()
