--- app/src/main/java/io/github/lens0021/teogeul/KOKR/layout/LayoutAlphabet.kt
@@ -1,98 +1,0 @@
-package io.github.lens0021.teogeul.KOKR.layout
-
-object LayoutAlphabet {
-    @JvmField
-    val CONVERT_ENGLISH_DVORAK =
-        arrayOf(
-            intArrayOf(49, 0x31, 0x21),
-            intArrayOf(50, 0x32, 0x40),
-            intArrayOf(51, 0x33, 0x23),
-            intArrayOf(52, 0x34, 0x24),
-            intArrayOf(53, 0x35, 0x25),
-            intArrayOf(54, 0x36, 0x5e),
-            intArrayOf(55, 0x37, 0x26),
-            intArrayOf(56, 0x38, 0x2a),
-            intArrayOf(57, 0x39, 0x28),
-            intArrayOf(48, 0x30, 0x29),
-            intArrayOf(113, 0x27, 0x22),
-            intArrayOf(119, 0x2c, 0x3c),
-            intArrayOf(101, 0x2e, 0x3e),
-            intArrayOf(114, 112, 80),
-            intArrayOf(116, 121, 89),
-            intArrayOf(121, 102, 70),
-            intArrayOf(117, 103, 71),
-            intArrayOf(105, 99, 67),
-            intArrayOf(111, 114, 82),
-            intArrayOf(112, 108, 76),
-            intArrayOf(97, 97, 65),
-            intArrayOf(115, 111, 79),
-            intArrayOf(100, 101, 69),
-            intArrayOf(102, 117, 85),
-            intArrayOf(103, 105, 73),
-            intArrayOf(104, 100, 68),
-            intArrayOf(106, 104, 72),
-            intArrayOf(107, 116, 84),
-            intArrayOf(108, 110, 78),
-            intArrayOf(0x3b, 115, 83),
-            intArrayOf(0x27, 0x2d, 0x5f),
-            intArrayOf(122, 0x3b, 0x3a),
-            intArrayOf(120, 113, 81),
-            intArrayOf(99, 106, 74),
-            intArrayOf(118, 107, 75),
-            intArrayOf(98, 120, 88),
-            intArrayOf(110, 98, 66),
-            intArrayOf(109, 109, 77),
-            intArrayOf(0x2c, 119, 87),
-            intArrayOf(0x2e, 118, 86),
-            intArrayOf(0x2f, 122, 90),
-            intArrayOf(0x2d, 0x5b, 0x7b),
-            intArrayOf(0x3d, 0x5d, 0x7d),
-            intArrayOf(0x5b, 0x2f, 0x3f),
-            intArrayOf(0x5d, 0x3d, 0x2b),
-        )
-
-    @JvmField
-    val CONVERT_ENGLISH_COLEMAK =
-        arrayOf(
-            intArrayOf(49, 0x31, 0x21),
-            intArrayOf(50, 0x32, 0x40),
-            intArrayOf(51, 0x33, 0x23),
-            intArrayOf(52, 0x34, 0x24),
-            intArrayOf(53, 0x35, 0x25),
-            intArrayOf(54, 0x36, 0x5e),
-            intArrayOf(55, 0x37, 0x26),
-            intArrayOf(56, 0x38, 0x2a),
-            intArrayOf(57, 0x39, 0x28),
-            intArrayOf(48, 0x30, 0x29),
-            intArrayOf(113, 113, 81),
-            intArrayOf(119, 119, 87),
-            intArrayOf(101, 102, 70),
-            intArrayOf(114, 112, 80),
-            intArrayOf(116, 103, 71),
-            intArrayOf(121, 106, 74),
-            intArrayOf(117, 108, 76),
-            intArrayOf(105, 117, 85),
-            intArrayOf(111, 121, 89),
-            intArrayOf(112, 0x3b, 0x3a),
-            intArrayOf(97, 97, 65),
-            intArrayOf(115, 114, 82),
-            intArrayOf(100, 115, 83),
-            intArrayOf(102, 116, 84),
-            intArrayOf(103, 100, 68),
-            intArrayOf(104, 104, 72),
-            intArrayOf(106, 110, 78),
-            intArrayOf(107, 101, 69),
-            intArrayOf(108, 105, 73),
-            intArrayOf(0x3b, 111, 79),
-            intArrayOf(122, 122, 90),
-            intArrayOf(120, 120, 88),
-            intArrayOf(99, 99, 67),
-            intArrayOf(118, 118, 86),
-            intArrayOf(98, 98, 66),
-            intArrayOf(110, 107, 75),
-            intArrayOf(109, 109, 77),
-            intArrayOf(0x2c, 0x2c, 0x3c),
-            intArrayOf(0x2e, 0x2e, 0x3e),
-            intArrayOf(0x2f, 0x2f, 0x3f),
-        )
-}
