--- app/src/main/java/io/github/lens0021/teogeul/KOKR/layout/LayoutDev.kt
@@ -1,113 +1,0 @@
-package io.github.lens0021.teogeul.KOKR.layout
-
-import io.github.lens0021.teogeul.KOKR.HangulEngine
-
-object LayoutDev {
-    @JvmField
-    val JAMO_NEBUL_1969 =
-        arrayOf(
-            intArrayOf(49, 0x31, 0x21),
-            intArrayOf(50, 0x32, 0x22),
-            intArrayOf(51, 0x33, 0xffe6),
-            intArrayOf(52, 0x34, 0x24),
-            intArrayOf(53, 0x35, 0x25),
-            intArrayOf(54, 0x36, 0x5f),
-            intArrayOf(55, 0x37, 0x3a),
-            intArrayOf(56, 0x38, 0x27),
-            intArrayOf(57, 0x39, 0x28),
-            intArrayOf(48, 0x30, 0x29),
-            intArrayOf(113, 0x1107, 0x1108),
-            intArrayOf(119, 0x110c, 0x110d),
-            intArrayOf(101, 0x1103, 0x1104),
-            intArrayOf(114, 0x1100, 0x1101),
-            intArrayOf(116, 0x1109, 0x110a),
-            intArrayOf(121, 0x1169, 0x116d),
-            intArrayOf(117, 0x1165, 0x1167),
-            intArrayOf(105, 0x1161, 0x1163),
-            intArrayOf(111, 0x1175, 0x1162),
-            intArrayOf(112, 0x1166, 0x3f),
-            intArrayOf(0x5b, 0x11a8, 0x11bd),
-            intArrayOf(97, 0x1106, 0x110f),
-            intArrayOf(115, 0x1102, 0x1110),
-            intArrayOf(100, 0x110b, 0x110e),
-            intArrayOf(102, 0x1105, 0x1111),
-            intArrayOf(103, 0x1112, 0x2f),
-            intArrayOf(104, 0x021169, 0x02116d),
-            intArrayOf(106, 0x021165, 0x021167),
-            intArrayOf(107, 0x021161, 0x021163),
-            intArrayOf(108, 0x021175, 0x021162),
-            intArrayOf(59, 0x021166, 0x021168),
-            intArrayOf(39, 0x11b8, 0x11ae),
-            intArrayOf(122, 0x11b7, 0x11c2),
-            intArrayOf(120, 0x11ab, 0x11c0),
-            intArrayOf(99, 0x11bc, 0x11be),
-            intArrayOf(118, 0x11af, 0x11c1),
-            intArrayOf(98, 0x02116e, 0x021172),
-            intArrayOf(110, 0x116e, 0x1172),
-            intArrayOf(109, 0x1173, 0x2e),
-            intArrayOf(44, 0x021173, 0x2c),
-            intArrayOf(46, 0x11ba, 0x11b9),
-            intArrayOf(47, 0x11bb, 0x11ad),
-            intArrayOf(0x2d, 0x2d, 0x2715),
-            intArrayOf(0x3d, 0x3d, 0x2b),
-            intArrayOf(128, 0x2e, 0x2c),
-        )
-
-    @JvmField
-    val COMB_NEBUL_1969 =
-        arrayOf(
-            intArrayOf(0x1169, 0x1161, 0x116a),
-            intArrayOf(0x1169, 0x1162, 0x116b),
-            intArrayOf(0x1169, 0x1175, 0x116c),
-            intArrayOf(0x116e, 0x1165, 0x116f),
-            intArrayOf(0x116e, 0x1166, 0x1170),
-            intArrayOf(0x116e, 0x1175, 0x1171),
-            intArrayOf(0x1173, 0x1175, 0x1174),
-            intArrayOf(0x1169, 0x021161, 0x02116a),
-            intArrayOf(0x1169, 0x021162, 0x02116b),
-            intArrayOf(0x1169, 0x021175, 0x02116c),
-            intArrayOf(0x116e, 0x021165, 0x02116f),
-            intArrayOf(0x116e, 0x021166, 0x021170),
-            intArrayOf(0x116e, 0x021175, 0x021171),
-            intArrayOf(0x1173, 0x021175, 0x021174),
-            intArrayOf(0x11a8, 0x11a8, 0x11a9),
-            intArrayOf(0x11a8, 0x11ba, 0x11aa),
-            intArrayOf(0x11ab, 0x11bd, 0x11ac),
-            intArrayOf(0x11ab, 0x11c2, 0x11ad),
-            intArrayOf(0x11af, 0x11a8, 0x11b0),
-            intArrayOf(0x11af, 0x11b7, 0x11b1),
-            intArrayOf(0x11af, 0x11b8, 0x11b2),
-            intArrayOf(0x11af, 0x11ba, 0x11b3),
-            intArrayOf(0x11af, 0x11c0, 0x11b4),
-            intArrayOf(0x11af, 0x11c1, 0x11b5),
-            intArrayOf(0x11af, 0x11c2, 0x11b6),
-            intArrayOf(0x11b8, 0x11ba, 0x11b9),
-            intArrayOf(0x11ba, 0x11ba, 0x11bb),
-        )
-
-    @JvmField
-    val VIRTUAL_NEBUL_1969 =
-        arrayOf(
-            intArrayOf(HangulEngine.VIRTUAL_JUNG, 0x021161, 0x1161),
-            intArrayOf(HangulEngine.VIRTUAL_JUNG, 0x021162, 0x1162),
-            intArrayOf(HangulEngine.VIRTUAL_JUNG, 0x021163, 0x1163),
-            intArrayOf(HangulEngine.VIRTUAL_JUNG, 0x021164, 0x1164),
-            intArrayOf(HangulEngine.VIRTUAL_JUNG, 0x021165, 0x1165),
-            intArrayOf(HangulEngine.VIRTUAL_JUNG, 0x021166, 0x1166),
-            intArrayOf(HangulEngine.VIRTUAL_JUNG, 0x021167, 0x1167),
-            intArrayOf(HangulEngine.VIRTUAL_JUNG, 0x021168, 0x1168),
-            intArrayOf(HangulEngine.VIRTUAL_JUNG, 0x021169, 0x1169),
-            intArrayOf(HangulEngine.VIRTUAL_JUNG, 0x02116a, 0x116a),
-            intArrayOf(HangulEngine.VIRTUAL_JUNG, 0x02116b, 0x116b),
-            intArrayOf(HangulEngine.VIRTUAL_JUNG, 0x02116c, 0x116c),
-            intArrayOf(HangulEngine.VIRTUAL_JUNG, 0x02116d, 0x116d),
-            intArrayOf(HangulEngine.VIRTUAL_JUNG, 0x02116e, 0x116e),
-            intArrayOf(HangulEngine.VIRTUAL_JUNG, 0x02116f, 0x116f),
-            intArrayOf(HangulEngine.VIRTUAL_JUNG, 0x021170, 0x1170),
-            intArrayOf(HangulEngine.VIRTUAL_JUNG, 0x021171, 0x1171),
-            intArrayOf(HangulEngine.VIRTUAL_JUNG, 0x021172, 0x1172),
-            intArrayOf(HangulEngine.VIRTUAL_JUNG, 0x021173, 0x1173),
-            intArrayOf(HangulEngine.VIRTUAL_JUNG, 0x021174, 0x1174),
-            intArrayOf(HangulEngine.VIRTUAL_JUNG, 0x021175, 0x1175),
-        )
-}
