--- app/src/main/java/io/github/lens0021/teogeul/KOKR/layout/LayoutDubul.kt
@@ -1,194 +1,0 @@
-package io.github.lens0021.teogeul.KOKR.layout
-
-object LayoutDubul {
-    @JvmField
-    val JAMO_DUBUL_STANDARD =
-        arrayOf(
-            intArrayOf(49, 0x31, 0x21),
-            intArrayOf(50, 0x32, 0x40),
-            intArrayOf(51, 0x33, 0x23),
-            intArrayOf(52, 0x34, 0x24),
-            intArrayOf(53, 0x35, 0x25),
-            intArrayOf(54, 0x36, 0x5e),
-            intArrayOf(55, 0x37, 0x26),
-            intArrayOf(56, 0x38, 0x2a),
-            intArrayOf(57, 0x39, 0x28),
-            intArrayOf(48, 0x30, 0x29),
-            intArrayOf(113, 0x3142, 0x3143),
-            intArrayOf(119, 0x3148, 0x3149),
-            intArrayOf(101, 0x3137, 0x3138),
-            intArrayOf(114, 0x3131, 0x3132),
-            intArrayOf(116, 0x3145, 0x3146),
-            intArrayOf(121, 0x315b, 0x315b),
-            intArrayOf(117, 0x3155, 0x3155),
-            intArrayOf(105, 0x3151, 0x3151),
-            intArrayOf(111, 0x3150, 0x3152),
-            intArrayOf(112, 0x3154, 0x3156),
-            intArrayOf(97, 0x3141, 0x3141),
-            intArrayOf(115, 0x3134, 0x3134),
-            intArrayOf(100, 0x3147, 0x3147),
-            intArrayOf(102, 0x3139, 0x3139),
-            intArrayOf(103, 0x314e, 0x314e),
-            intArrayOf(104, 0x3157, 0x3157),
-            intArrayOf(106, 0x3153, 0x3153),
-            intArrayOf(107, 0x314f, 0x314f),
-            intArrayOf(108, 0x3163, 0x3163),
-            intArrayOf(122, 0x314b, 0x314b),
-            intArrayOf(120, 0x314c, 0x314c),
-            intArrayOf(99, 0x314a, 0x314a),
-            intArrayOf(118, 0x314d, 0x314d),
-            intArrayOf(98, 0x3160, 0x3160),
-            intArrayOf(110, 0x315c, 0x315c),
-            intArrayOf(109, 0x3161, 0x3161),
-        )
-
-    @JvmField
-    val COMB_DUBUL_STANDARD =
-        arrayOf(
-            intArrayOf(0x1169, 0x1161, 0x116a),
-            intArrayOf(0x1169, 0x1162, 0x116b),
-            intArrayOf(0x1169, 0x1175, 0x116c),
-            intArrayOf(0x116e, 0x1165, 0x116f),
-            intArrayOf(0x116e, 0x1166, 0x1170),
-            intArrayOf(0x116e, 0x1175, 0x1171),
-            intArrayOf(0x1173, 0x1175, 0x1174),
-            intArrayOf(0x11a8, 0x11ba, 0x11aa),
-            intArrayOf(0x11ab, 0x11bd, 0x11ac),
-            intArrayOf(0x11ab, 0x11c2, 0x11ad),
-            intArrayOf(0x11af, 0x11a8, 0x11b0),
-            intArrayOf(0x11af, 0x11b7, 0x11b1),
-            intArrayOf(0x11af, 0x11b8, 0x11b2),
-            intArrayOf(0x11af, 0x11ba, 0x11b3),
-            intArrayOf(0x11af, 0x11c0, 0x11b4),
-            intArrayOf(0x11af, 0x11c1, 0x11b5),
-            intArrayOf(0x11af, 0x11c2, 0x11b6),
-            intArrayOf(0x11b8, 0x11ba, 0x11b9),
-        )
-
-    @JvmField
-    val JAMO_DUBUL_NK =
-        arrayOf(
-            intArrayOf(113, 0x3142, 0x3143),
-            intArrayOf(119, 0x3141, 0x3141),
-            intArrayOf(101, 0x3137, 0x3138),
-            intArrayOf(114, 0x3139, 0x3139),
-            intArrayOf(116, 0x314e, 0x314e),
-            intArrayOf(121, 0x3155, 0x3155),
-            intArrayOf(117, 0x315c, 0x315c),
-            intArrayOf(105, 0x3153, 0x3153),
-            intArrayOf(111, 0x3150, 0x3152),
-            intArrayOf(112, 0x3154, 0x3156),
-            intArrayOf(97, 0x3148, 0x3149),
-            intArrayOf(115, 0x3131, 0x3132),
-            intArrayOf(100, 0x3147, 0x3147),
-            intArrayOf(102, 0x3134, 0x3134),
-            intArrayOf(103, 0x3145, 0x3146),
-            intArrayOf(104, 0x3157, 0x3157),
-            intArrayOf(106, 0x314f, 0x314f),
-            intArrayOf(107, 0x3163, 0x3163),
-            intArrayOf(108, 0x3161, 0x3161),
-            intArrayOf(122, 0x314b, 0x314b),
-            intArrayOf(120, 0x314c, 0x314c),
-            intArrayOf(99, 0x314a, 0x314a),
-            intArrayOf(118, 0x314d, 0x314d),
-            intArrayOf(98, 0x3160, 0x3160),
-            intArrayOf(110, 0x315b, 0x315b),
-            intArrayOf(109, 0x3151, 0x3151),
-        )
-
-    @JvmField
-    val JAMO_DUBUL_YET =
-        arrayOf(
-            intArrayOf(113, 0x3142, 0x3143),
-            intArrayOf(119, 0x3148, 0x3149),
-            intArrayOf(101, 0x3137, 0x3138),
-            intArrayOf(114, 0x3131, 0x3132),
-            intArrayOf(116, 0x3145, 0x3146),
-            intArrayOf(121, 0x315b, 0x315b),
-            intArrayOf(117, 0x3155, 0x3155),
-            intArrayOf(105, 0x3151, 0x3151),
-            intArrayOf(111, 0x3150, 0x3152),
-            intArrayOf(112, 0x3154, 0x3156),
-            intArrayOf(97, 0x3141, 0x317f),
-            intArrayOf(115, 0x3134, 0x3136),
-            intArrayOf(100, 0x3147, 0x3181),
-            intArrayOf(102, 0x3139, 0x3140),
-            intArrayOf(103, 0x314e, 0x3186),
-            intArrayOf(104, 0x3157, 0x1183),
-            intArrayOf(106, 0x3153, 0x115f),
-            intArrayOf(107, 0x314f, 0x318d),
-            intArrayOf(108, 0x3163, 0x318c),
-            intArrayOf(122, 0x314b, 0x113c),
-            intArrayOf(120, 0x314c, 0x113e),
-            intArrayOf(99, 0x314a, 0x114e),
-            intArrayOf(118, 0x314d, 0x1150),
-            intArrayOf(98, 0x3160, 0x1154),
-            intArrayOf(110, 0x315c, 0x1155),
-            intArrayOf(109, 0x3161, 0x3161),
-        )
-
-    @JvmField
-    val JAMO_DUBUL_DANMOEUM_GOOGLE =
-        arrayOf(
-            intArrayOf(113, 0x3142),
-            intArrayOf(119, 0x3148),
-            intArrayOf(101, 0x3137),
-            intArrayOf(114, 0x3131),
-            intArrayOf(116, 0x3145),
-            intArrayOf(105, 0x3157),
-            intArrayOf(111, 0x3150),
-            intArrayOf(112, 0x3154),
-            intArrayOf(97, 0x3141),
-            intArrayOf(115, 0x3134),
-            intArrayOf(100, 0x3147),
-            intArrayOf(102, 0x3139),
-            intArrayOf(104, 0x314e),
-            intArrayOf(106, 0x3153),
-            intArrayOf(107, 0x314f),
-            intArrayOf(108, 0x3163),
-            intArrayOf(120, 0x314b),
-            intArrayOf(99, 0x314c),
-            intArrayOf(118, 0x314a),
-            intArrayOf(98, 0x314d),
-            intArrayOf(110, 0x315c),
-            intArrayOf(109, 0x3161),
-        )
-
-    @JvmField
-    val COMB_DUBUL_DANMOEUM_GOOGLE =
-        arrayOf(
-            intArrayOf(0x1100, 0x1100, 0x1101),
-            intArrayOf(0x1103, 0x1103, 0x1104),
-            intArrayOf(0x1107, 0x1107, 0x1108),
-            intArrayOf(0x1109, 0x1109, 0x110a),
-            intArrayOf(0x110c, 0x110c, 0x110d),
-            intArrayOf(0x1161, 0x1161, 0x1163),
-            intArrayOf(0x1162, 0x1162, 0x1164),
-            intArrayOf(0x1163, 0x1175, 0x1164),
-            intArrayOf(0x1165, 0x1165, 0x1167),
-            intArrayOf(0x1167, 0x1175, 0x1168),
-            intArrayOf(0x1166, 0x1166, 0x1168),
-            intArrayOf(0x1169, 0x1161, 0x116a),
-            intArrayOf(0x1169, 0x1162, 0x116b),
-            intArrayOf(0x1169, 0x1175, 0x116c),
-            intArrayOf(0x1169, 0x1169, 0x116d),
-            intArrayOf(0x116e, 0x1165, 0x116f),
-            intArrayOf(0x116e, 0x1166, 0x1170),
-            intArrayOf(0x116e, 0x1175, 0x1171),
-            intArrayOf(0x116e, 0x116e, 0x1172),
-            intArrayOf(0x1173, 0x1175, 0x1174),
-            intArrayOf(0x11a8, 0x11a8, 0x11a9),
-            intArrayOf(0x11a8, 0x11ba, 0x11aa),
-            intArrayOf(0x11ab, 0x11bd, 0x11ac),
-            intArrayOf(0x11ab, 0x11c2, 0x11ad),
-            intArrayOf(0x11af, 0x11a8, 0x11b0),
-            intArrayOf(0x11af, 0x11b7, 0x11b1),
-            intArrayOf(0x11af, 0x11b8, 0x11b2),
-            intArrayOf(0x11af, 0x11ba, 0x11b3),
-            intArrayOf(0x11af, 0x11c0, 0x11b4),
-            intArrayOf(0x11af, 0x11c1, 0x11b5),
-            intArrayOf(0x11af, 0x11c2, 0x11b6),
-            intArrayOf(0x11b8, 0x11ba, 0x11b9),
-            intArrayOf(0x11ba, 0x11ba, 0x11bb),
-        )
-}
