--- app/src/main/java/io/github/lens0021/teogeul/KOKR/layout/LayoutShinSebul.kt
@@ -1,1026 +1,0 @@
-package io.github.lens0021.teogeul.KOKR.layout
-
-object LayoutShinSebul {
-    @JvmField
-    val JAMO_SEBUL_3_2015M_CHOJONG =
-        arrayOf(
-            intArrayOf(49, 0x11ae, 0x21),
-            intArrayOf(50, 0x11bb, 0x40),
-            intArrayOf(51, 0x11b8, 0x23),
-            intArrayOf(52, 0x116d, 0x24),
-            intArrayOf(53, 0x1172, 0x25),
-            intArrayOf(54, 0x1163, 0x5e),
-            intArrayOf(55, 0x1168, 0x26),
-            intArrayOf(56, 0x1174, 0x2a),
-            intArrayOf(57, 0x01116e, 0x28),
-            intArrayOf(48, 0x110f, 0x29),
-            intArrayOf(113, 0x11ba, 0x11a9), // q
-            intArrayOf(119, 0x11af, 0x11b0), // w
-            intArrayOf(101, 0x11bd, 0x1167), // e
-            intArrayOf(114, 0x11be, 0x1162), // r
-            intArrayOf(116, 0x1165, 0x1164), // t
-            intArrayOf(121, 0x1105, 0x35), // y
-            intArrayOf(117, 0x1103, 0x36), // u
-            intArrayOf(105, 0x1106, 0x37), // i
-            intArrayOf(111, 0x110e, 0x38), // o
-            intArrayOf(112, 0x1111, 0x39), // p
-            intArrayOf(97, 0x11bc, 0x11b4), // a
-            intArrayOf(115, 0x11ab, 0x11ad), // s
-            intArrayOf(100, 0x11c2, 0x1175), // d
-            intArrayOf(102, 0x11c1, 0x1161), // f
-            intArrayOf(103, 0x1173, 0x3a), // g
-            intArrayOf(104, 0x1102, 0x30), // h
-            intArrayOf(106, 0x110b, 0x31), // j
-            intArrayOf(107, 0x1100, 0x32), // k
-            intArrayOf(108, 0x110c, 0x33), // l
-            intArrayOf(59, 0x1107, 0x34),
-            intArrayOf(39, 0x1110, 0x2f),
-            intArrayOf(122, 0x11b7, 0x11b3), // z
-            intArrayOf(120, 0x11a8, 0x11b9), // x
-            intArrayOf(99, 0x11c0, 0x1166), // c
-            intArrayOf(118, 0x11bf, 0x1169), // v
-            intArrayOf(98, 0x116e, 0x3b), // b
-            intArrayOf(110, 0x1109, 0x27), // n
-            intArrayOf(109, 0x1112, 0x22), // m
-            intArrayOf(44, 0x2c, 0x3c),
-            intArrayOf(46, 0x2e, 0x3e),
-            intArrayOf(47, 0x011169, 0x3f),
-            intArrayOf(128, 0x2e, 0x2c),
-        )
-
-    @JvmField
-    val JAMO_SEBUL_3_2015M_CHOJUNG =
-        arrayOf(
-            intArrayOf(49, 0x11ae, 0x21),
-            intArrayOf(50, 0x11bb, 0x40),
-            intArrayOf(51, 0x11b8, 0x23),
-            intArrayOf(52, 0x116d, 0x24),
-            intArrayOf(53, 0x1172, 0x25),
-            intArrayOf(54, 0x1163, 0x5e),
-            intArrayOf(55, 0x1168, 0x26),
-            intArrayOf(56, 0x1174, 0x2a),
-            intArrayOf(57, 0x01116e, 0x28),
-            intArrayOf(48, 0x110f, 0x29),
-            intArrayOf(113, 0x11ba, 0x11a9), // q
-            intArrayOf(119, 0x11af, 0x11b0), // w
-            intArrayOf(101, 0x1167, 0x11bd), // e
-            intArrayOf(114, 0x1162, 0x11be), // r
-            intArrayOf(116, 0x1165, 0x1164), // t
-            intArrayOf(121, 0x1105, 0x0035), // y
-            intArrayOf(117, 0x1103, 0x0036), // u
-            intArrayOf(105, 0x1106, 0x0037), // i
-            intArrayOf(111, 0x110e, 0x0038), // o
-            intArrayOf(112, 0x1111, 0x0039), // p
-            intArrayOf(97, 0x11bc, 0x11b4), // a
-            intArrayOf(115, 0x11ab, 0x11ad), // s
-            intArrayOf(100, 0x1175, 0x11c2), // d
-            intArrayOf(102, 0x1161, 0x11c1), // f
-            intArrayOf(103, 0x1173, 0x3a), // g
-            intArrayOf(104, 0x1102, 0x30), // h
-            intArrayOf(106, 0x110b, 0x31), // j
-            intArrayOf(107, 0x1100, 0x32), // k
-            intArrayOf(108, 0x110c, 0x33), // l
-            intArrayOf(59, 0x1107, 0x34),
-            intArrayOf(39, 0x1110, 0x2f),
-            intArrayOf(122, 0x11b7, 0x11b3), // z
-            intArrayOf(120, 0x11a8, 0x11b9), // x
-            intArrayOf(99, 0x1166, 0x11c0), // c
-            intArrayOf(118, 0x1169, 0x11bf), // v
-            intArrayOf(98, 0x116e, 0x3b), // b
-            intArrayOf(110, 0x1109, 0x27), // n
-            intArrayOf(109, 0x1112, 0x22), // m
-            intArrayOf(44, 0x2c, 0x3c),
-            intArrayOf(46, 0x2e, 0x3e),
-            intArrayOf(47, 0x011169, 0x3f),
-            intArrayOf(128, 0x2e, 0x2c),
-        )
-
-    // 윗글쇠 자리에 겹받침이 들어간 배열.
-    @JvmField
-    val JAMO_SEBUL_3_2015M_CHOJONG_D =
-        arrayOf(
-            intArrayOf(49, 0x11ae, 0x21),
-            intArrayOf(50, 0x11bb, 0x40),
-            intArrayOf(51, 0x11b8, 0x23),
-            intArrayOf(52, 0x116d, 0x24),
-            intArrayOf(53, 0x1172, 0x25),
-            intArrayOf(54, 0x1163, 0x5e),
-            intArrayOf(55, 0x1168, 0x26),
-            intArrayOf(56, 0x1174, 0x2a),
-            intArrayOf(57, 0x01116e, 0x28),
-            intArrayOf(48, 0x110f, 0x29),
-            intArrayOf(113, 0x11ba, 0x11a9), // q
-            intArrayOf(119, 0x11af, 0x11b0), // w
-            intArrayOf(101, 0x11bd, 0x11ac), // e
-            intArrayOf(114, 0x11be, 0x11b6), // r
-            intArrayOf(116, 0x1165, 0x1164), // t
-            intArrayOf(121, 0x1105, 0x35), // y
-            intArrayOf(117, 0x1103, 0x36), // u
-            intArrayOf(105, 0x1106, 0x37), // i
-            intArrayOf(111, 0x110e, 0x38), // o
-            intArrayOf(112, 0x1111, 0x39), // p
-            intArrayOf(97, 0x11bc, 0x11b4), // a
-            intArrayOf(115, 0x11ab, 0x11ad), // s
-            intArrayOf(100, 0x11c2, 0x11b2), // d
-            intArrayOf(102, 0x11c1, 0x11b1), // f
-            intArrayOf(103, 0x1173, 0x3a), // g
-            intArrayOf(104, 0x1102, 0x30), // h
-            intArrayOf(106, 0x110b, 0x31), // j
-            intArrayOf(107, 0x1100, 0x32), // k
-            intArrayOf(108, 0x110c, 0x33), // l
-            intArrayOf(59, 0x1107, 0x34),
-            intArrayOf(39, 0x1110, 0x2f),
-            intArrayOf(122, 0x11b7, 0x11b3), // z
-            intArrayOf(120, 0x11a8, 0x11b9), // x
-            intArrayOf(99, 0x11c0, 0x116b5), // c
-            intArrayOf(118, 0x11bf, 0x11aa), // v
-            intArrayOf(98, 0x116e, 0x3b), // b
-            intArrayOf(110, 0x1109, 0x27), // n
-            intArrayOf(109, 0x1112, 0x22), // m
-            intArrayOf(44, 0x2c, 0x3c),
-            intArrayOf(46, 0x2e, 0x3e),
-            intArrayOf(47, 0x011169, 0x3f),
-            intArrayOf(128, 0x2e, 0x2c),
-        )
-
-    @JvmField
-    val JAMOSET_SEBUL_3_2015M =
-        arrayOf(
-            JAMO_SEBUL_3_2015M_CHOJUNG,
-            JAMO_SEBUL_3_2015M_CHOJUNG,
-            JAMO_SEBUL_3_2015M_CHOJONG_D,
-            JAMO_SEBUL_3_2015M_CHOJONG,
-        )
-
-    @JvmField
-    val JAMO_SEBUL_3_2015_CHOJONG =
-        arrayOf(
-            intArrayOf(49, 0x11ae, 0x21),
-            intArrayOf(50, 0x11bb, 0x40),
-            intArrayOf(51, 0x11b8, 0x23),
-            intArrayOf(52, 0x116d, 0x24),
-            intArrayOf(53, 0x1172, 0x25),
-            intArrayOf(54, 0x1163, 0x5e),
-            intArrayOf(55, 0x1168, 0x26),
-            intArrayOf(56, 0x1174, 0x2a),
-            intArrayOf(57, 0x01116e, 0x28),
-            intArrayOf(48, 0x110f, 0x29),
-            intArrayOf(113, 0x11ba, 0x11a9), // q
-            intArrayOf(119, 0x11af, 0x11b0), // w
-            intArrayOf(101, 0x11bd, 0x1167), // e
-            intArrayOf(114, 0x11be, 0x1165), // r
-            intArrayOf(116, 0x1162, 0x1164), // t
-            intArrayOf(121, 0x1105, 0x35), // y
-            intArrayOf(117, 0x1103, 0x36), // u
-            intArrayOf(105, 0x1106, 0x37), // i
-            intArrayOf(111, 0x110e, 0x38), // o
-            intArrayOf(112, 0x1111, 0x39), // p
-            intArrayOf(97, 0x11bc, 0x11b4), // a
-            intArrayOf(115, 0x11ab, 0x11ad), // s
-            intArrayOf(100, 0x11c2, 0x1175), // d
-            intArrayOf(102, 0x11c1, 0x1161), // f
-            intArrayOf(103, 0x1173, 0x3a), // g
-            intArrayOf(104, 0x1102, 0x30), // h
-            intArrayOf(106, 0x110b, 0x31), // j
-            intArrayOf(107, 0x1100, 0x32), // k
-            intArrayOf(108, 0x110c, 0x33), // l
-            intArrayOf(59, 0x1107, 0x34),
-            intArrayOf(39, 0x1110, 0x2f),
-            intArrayOf(122, 0x11b7, 0x11b3), // z
-            intArrayOf(120, 0x11a8, 0x11b9), // x
-            intArrayOf(99, 0x11c0, 0x1166), // c
-            intArrayOf(118, 0x11bf, 0x1169), // v
-            intArrayOf(98, 0x116e, 0x3b), // b
-            intArrayOf(110, 0x1109, 0x27), // n
-            intArrayOf(109, 0x1112, 0x22), // m
-            intArrayOf(44, 0x2c, 0x3c),
-            intArrayOf(46, 0x2e, 0x3e),
-            intArrayOf(47, 0x011169, 0x3f),
-            intArrayOf(128, 0x2e, 0x2c),
-        )
-
-    @JvmField
-    val JAMO_SEBUL_3_2015_CHOJUNG =
-        arrayOf(
-            intArrayOf(49, 0x11ae, 0x21),
-            intArrayOf(50, 0x11bb, 0x40),
-            intArrayOf(51, 0x11b8, 0x23),
-            intArrayOf(52, 0x116d, 0x24),
-            intArrayOf(53, 0x1172, 0x25),
-            intArrayOf(54, 0x1163, 0x5e),
-            intArrayOf(55, 0x1168, 0x26),
-            intArrayOf(56, 0x1174, 0x2a),
-            intArrayOf(57, 0x01116e, 0x28),
-            intArrayOf(48, 0x110f, 0x29),
-            intArrayOf(113, 0x11ba, 0x11a9), // q
-            intArrayOf(119, 0x11af, 0x11b0), // w
-            intArrayOf(101, 0x1167, 0x11bd), // e
-            intArrayOf(114, 0x1165, 0x11be), // r
-            intArrayOf(116, 0x1162, 0x1164), // t
-            intArrayOf(121, 0x1105, 0x0035), // y
-            intArrayOf(117, 0x1103, 0x0036), // u
-            intArrayOf(105, 0x1106, 0x0037), // i
-            intArrayOf(111, 0x110e, 0x0038), // o
-            intArrayOf(112, 0x1111, 0x0039), // p
-            intArrayOf(97, 0x11bc, 0x11b4), // a
-            intArrayOf(115, 0x11ab, 0x11ad), // s
-            intArrayOf(100, 0x1175, 0x11c2), // d
-            intArrayOf(102, 0x1161, 0x11c1), // f
-            intArrayOf(103, 0x1173, 0x3a), // g
-            intArrayOf(104, 0x1102, 0x30), // h
-            intArrayOf(106, 0x110b, 0x31), // j
-            intArrayOf(107, 0x1100, 0x32), // k
-            intArrayOf(108, 0x110c, 0x33), // l
-            intArrayOf(59, 0x1107, 0x34),
-            intArrayOf(39, 0x1110, 0x2f),
-            intArrayOf(122, 0x11b7, 0x11b3), // z
-            intArrayOf(120, 0x11a8, 0x11b9), // x
-            intArrayOf(99, 0x1166, 0x11c0), // c
-            intArrayOf(118, 0x1169, 0x11bf), // v
-            intArrayOf(98, 0x116e, 0x3b), // b
-            intArrayOf(110, 0x1109, 0x27), // n
-            intArrayOf(109, 0x1112, 0x22), // m
-            intArrayOf(44, 0x2c, 0x3c),
-            intArrayOf(46, 0x2e, 0x3e),
-            intArrayOf(47, 0x011169, 0x3f),
-            intArrayOf(128, 0x2e, 0x2c),
-        )
-
-    // 윗글쇠에 겹받침이 있는 배열.
-    @JvmField
-    val JAMO_SEBUL_3_2015_CHOJONG_D =
-        arrayOf(
-            intArrayOf(49, 0x11ae, 0x21),
-            intArrayOf(50, 0x11bb, 0x40),
-            intArrayOf(51, 0x11b8, 0x23),
-            intArrayOf(52, 0x116d, 0x24),
-            intArrayOf(53, 0x1172, 0x25),
-            intArrayOf(54, 0x1163, 0x5e),
-            intArrayOf(55, 0x1168, 0x26),
-            intArrayOf(56, 0x1174, 0x2a),
-            intArrayOf(57, 0x01116e, 0x28),
-            intArrayOf(48, 0x110f, 0x29),
-            intArrayOf(113, 0x11ba, 0x11a9), // q
-            intArrayOf(119, 0x11af, 0x11b0), // w
-            intArrayOf(101, 0x11bd, 0x11ac), // e
-            intArrayOf(114, 0x11be, 0x11b6), // r
-            intArrayOf(116, 0x1162, 0x1164), // t
-            intArrayOf(121, 0x1105, 0x35), // y
-            intArrayOf(117, 0x1103, 0x36), // u
-            intArrayOf(105, 0x1106, 0x37), // i
-            intArrayOf(111, 0x110e, 0x38), // o
-            intArrayOf(112, 0x1111, 0x39), // p
-            intArrayOf(97, 0x11bc, 0x11b4), // a
-            intArrayOf(115, 0x11ab, 0x11ad), // s
-            intArrayOf(100, 0x11c2, 0x11b2), // d
-            intArrayOf(102, 0x11c1, 0x11b1), // f
-            intArrayOf(103, 0x1173, 0x3a), // g
-            intArrayOf(104, 0x1102, 0x30), // h
-            intArrayOf(106, 0x110b, 0x31), // j
-            intArrayOf(107, 0x1100, 0x32), // k
-            intArrayOf(108, 0x110c, 0x33), // l
-            intArrayOf(59, 0x1107, 0x34),
-            intArrayOf(39, 0x1110, 0x2f),
-            intArrayOf(122, 0x11b7, 0x11b3), // z
-            intArrayOf(120, 0x11a8, 0x11b9), // x
-            intArrayOf(99, 0x11c0, 0x11b5), // c
-            intArrayOf(118, 0x11bf, 0x11aa), // v
-            intArrayOf(98, 0x116e, 0x3b), // b
-            intArrayOf(110, 0x1109, 0x27), // n
-            intArrayOf(109, 0x1112, 0x22), // m
-            intArrayOf(44, 0x2c, 0x3c),
-            intArrayOf(46, 0x2e, 0x3e),
-            intArrayOf(47, 0x011169, 0x3f),
-            intArrayOf(128, 0x2e, 0x2c),
-        )
-
-    @JvmField
-    val JAMOSET_SEBUL_3_2015 =
-        arrayOf(
-            JAMO_SEBUL_3_2015_CHOJUNG,
-            JAMO_SEBUL_3_2015_CHOJUNG,
-            JAMO_SEBUL_3_2015_CHOJONG_D,
-            JAMO_SEBUL_3_2015_CHOJONG,
-        )
-
-    @JvmField
-    val JAMO_SEBUL_3_2015Y =
-        arrayOf(
-            intArrayOf(49, 0x11c2, 0x11f9),
-            intArrayOf(50, 0x11bb, 0x40),
-            intArrayOf(51, 0x11b8, 0x23),
-            intArrayOf(52, 0x116d, 0x24),
-            intArrayOf(53, 0x1172, 0x25),
-            intArrayOf(54, 0x1163, 0x5e),
-            intArrayOf(55, 0x1168, 0x26),
-            intArrayOf(56, 0x1174, 0x2a),
-            intArrayOf(57, 0x116e, 0x28),
-            intArrayOf(48, 0x110f, 0x29),
-            intArrayOf(113, 0x11ba, 0x11eb), // q
-            intArrayOf(119, 0x11af, 0x11af), // w
-            intArrayOf(101, 0x1167, 0x11bd), // e
-            intArrayOf(114, 0x1165, 0x11be), // r
-            intArrayOf(116, 0x1162, 0x1164), // t
-            intArrayOf(121, 0x1105, 0x302f), // y
-            intArrayOf(117, 0x1103, 0x302e), // u
-            intArrayOf(105, 0x1106, 0x1154), // i
-            intArrayOf(111, 0x110e, 0x1155), // o
-            intArrayOf(112, 0x1111, 0x3b), // p
-            intArrayOf(129, 0x27, 0x22),
-            intArrayOf(91, 0x5b, 0x27),
-            intArrayOf(93, 0x5d, 0x22),
-            intArrayOf(97, 0x11bc, 0x11f0), // a
-            intArrayOf(115, 0x11ab, 0x11ab), // s
-            intArrayOf(100, 0x1175, 0x11ae), // d
-            intArrayOf(102, 0x1161, 0x11c1), // f
-            intArrayOf(103, 0x1173, 0x119e), // g
-            intArrayOf(104, 0x1102, 0xb7), // h
-            intArrayOf(106, 0x110b, 0x114c), // j
-            intArrayOf(107, 0x1100, 0x114e), // k
-            intArrayOf(108, 0x110c, 0x1150), // l
-            intArrayOf(59, 0x1107, 0x3a),
-            intArrayOf(39, 0x1110, 0x2f),
-            intArrayOf(122, 0x11b7, 0x11b7), // z
-            intArrayOf(120, 0x11a8, 0x11a8), // x
-            intArrayOf(99, 0x1166, 0x11c0), // c
-            intArrayOf(118, 0x1169, 0x11bf), // v
-            intArrayOf(98, 0x116e, 0x0021), // b
-            intArrayOf(110, 0x1109, 0x1140), // n
-            intArrayOf(109, 0x1112, 0x1159), // m
-            intArrayOf(44, 0x2c, 0x113c),
-            intArrayOf(46, 0x2e, 0x113e),
-            intArrayOf(47, 0x1169, 0x3f),
-            intArrayOf(128, 0x2e, 0x2c),
-        )
-
-    @JvmField
-    val COMB_SEBUL_3_2015 =
-        arrayOf(
-            intArrayOf(0x1100, 0x110b, 0x1101), // ㄱ + ㅇ = ㄲ
-            intArrayOf(0x1100, 0x110c, 0x110d), // ㄱ + ㅈ = ㅉ
-            intArrayOf(0x1103, 0x1106, 0x1104), // ㄷ + ㅁ = ㄸ
-            intArrayOf(0x1106, 0x1103, 0x1104), // ㅁ + ㄷ = ㄸ
-            intArrayOf(0x1107, 0x110c, 0x1108), // ㅂ + ㅈ = ㅃ
-            intArrayOf(0x1109, 0x1112, 0x110a), // ㅅ + ㅎ = ㅆ
-            intArrayOf(0x110b, 0x1100, 0x1101), // ㅇ + ㄱ = ㄲ
-            intArrayOf(0x110c, 0x1100, 0x110d), // ㅈ + ㄱ = ㅉ
-            intArrayOf(0x110c, 0x1107, 0x1108), // ㅈ + ㅂ = ㅃ
-            intArrayOf(0x1112, 0x1109, 0x110a), // ㅎ + ㅅ = ㅆ
-            intArrayOf(0x1162, 0x1162, 0x1164), // ㅐ + ㅐ = ㅒ
-            intArrayOf(0x1165, 0x1165, 0x1164), // ㅓ + ㅓ = ㅒ
-            intArrayOf(0x1175, 0x1162, 0x1164), // ㅣ + ㅐ = ㅒ
-            intArrayOf(0x11a8, 0x11af, 0x11b0), // ㄱ + ㄹ = ㄺ
-            intArrayOf(0x11a8, 0x11b7, 0x11a9), // ㄱ + ㅁ = ㄲ
-            intArrayOf(0x11b7, 0x11a8, 0x11a9), // ㅁ + ㄱ = ㄲ
-            intArrayOf(0x11b7, 0x11af, 0x11b1), // ㅁ + ㄹ = ㄻ
-            intArrayOf(0x11b8, 0x11af, 0x11b2), // ㅂ + ㄹ = ㄼ
-            intArrayOf(0x11ba, 0x11a8, 0x11aa), // ㅅ + ㄱ = ㄳ
-            intArrayOf(0x11ba, 0x11af, 0x11b3), // ㅅ + ㄹ = ㄽ
-            intArrayOf(0x11ba, 0x11b8, 0x11b9), // ㅅ + ㅂ = ㅄ
-            intArrayOf(0x11bd, 0x11ab, 0x11ac), // ㅈ + ㄴ = ㄵ
-            intArrayOf(0x11c0, 0x11af, 0x11b4), // ㅌ + ㄹ = ㄾ
-            intArrayOf(0x11c1, 0x11af, 0x11b5), // ㅍ + ㄹ = ㄿ
-            intArrayOf(0x11c2, 0x11ab, 0x11ad), // ㅎ + ㄴ = ㄶ
-            intArrayOf(0x11c2, 0x11af, 0x11b6), // ㅎ + ㄹ = ㅀ
-            intArrayOf(0x1100, 0x1100, 0x1101),	// ㄲ
-            intArrayOf(0x1103, 0x1103, 0x1104),	// ㄸ
-            intArrayOf(0x1107, 0x1107, 0x1108),	// ㅃ
-            intArrayOf(0x1109, 0x1109, 0x110a),	// ㅆ
-            intArrayOf(0x110c, 0x110c, 0x110d),	// ㅉ
-            intArrayOf(0x1169, 0x1161, 0x116a),	// ㅘ
-            intArrayOf(0x1169, 0x1162, 0x116b),	// ㅙ
-            intArrayOf(0x1169, 0x1175, 0x116c),	// ㅚ
-            intArrayOf(0x116e, 0x1165, 0x116f),	// ㅝ
-            intArrayOf(0x116e, 0x1166, 0x1170),	// ㅞ
-            intArrayOf(0x116e, 0x1175, 0x1171),	// ㅟ
-            // 가상 낱자 조합
-            intArrayOf(0x011169, 0x1161, 0x116a),	// ㅘ
-            intArrayOf(0x011169, 0x1162, 0x116b),	// ㅙ
-            intArrayOf(0x011169, 0x1175, 0x116c),	// ㅚ
-            intArrayOf(0x01116e, 0x1165, 0x116f),	// ㅝ
-            intArrayOf(0x01116e, 0x1166, 0x1170),	// ㅞ
-            intArrayOf(0x01116e, 0x1175, 0x1171),	// ㅟ
-            intArrayOf(0x1173, 0x1175, 0x1174),	// ㅢ
-            intArrayOf(0x11a8, 0x11a8, 0x11a9),	// ㄲ
-            intArrayOf(0x11a8, 0x11ba, 0x11aa),	// ㄳ
-            intArrayOf(0x11ab, 0x11bd, 0x11ac),	// ㄵ
-            intArrayOf(0x11ab, 0x11c2, 0x11ad),	// ㄶ
-            intArrayOf(0x11af, 0x11a8, 0x11b0),	// ㄺ
-            intArrayOf(0x11af, 0x11b7, 0x11b1),	// ㄻ
-            intArrayOf(0x11af, 0x11b8, 0x11b2),	// ㄼ
-            intArrayOf(0x11af, 0x11ba, 0x11b3),	// ㄽ
-            intArrayOf(0x11af, 0x11c0, 0x11b4),	// ㄾ
-            intArrayOf(0x11af, 0x11c1, 0x11b5),	// ㄿ
-            intArrayOf(0x11af, 0x11c2, 0x11b6),	// ㅀ
-            intArrayOf(0x11b8, 0x11ba, 0x11b9),	// ㅄ
-            intArrayOf(0x11ba, 0x11ba, 0x11bb),	// ㅆ
-        )
-
-    @JvmField
-    val JAMO_SEBUL_3_P3_CHOJONG =
-        arrayOf(
-            intArrayOf(49, 0x11bf, 0x21),
-            intArrayOf(50, 0x11bb, 0x40),
-            intArrayOf(51, 0x11b8, 0x23),
-            intArrayOf(52, 0x116d, 0x24),
-            intArrayOf(53, 0x1172, 0x25),
-            intArrayOf(54, 0x1163, 0x5e),
-            intArrayOf(55, 0x1168, 0x26),
-            intArrayOf(56, 0x011173, 0x2a),
-            intArrayOf(57, 0x01116e, 0x28),
-            intArrayOf(48, 0x110f, 0x29),
-            intArrayOf(113, 0x11ba, 0x11b6), // q
-            intArrayOf(119, 0x11af, 0x11b0), // w
-            intArrayOf(101, 0x11c0, 0x1167), // e
-            intArrayOf(114, 0x11be, 0x1165), // r
-            intArrayOf(116, 0x1162, 0x1164), // t
-            intArrayOf(121, 0x1105, 0x2f), // y
-            intArrayOf(117, 0x1103, 0x37), // u
-            intArrayOf(105, 0x1106, 0x38), // i
-            intArrayOf(111, 0x110e, 0x39), // o
-            intArrayOf(112, 0x1111, 0x3b), // p
-            intArrayOf(97, 0x11bc, 0x11b9), // a
-            intArrayOf(115, 0x11ab, 0x11ad), // s
-            intArrayOf(100, 0x11c2, 0x1175), // d
-            intArrayOf(102, 0x11c1, 0x1161), // f
-            intArrayOf(103, 0x1173, 0x3c), // g
-            intArrayOf(104, 0x1102, 0x27), // h
-            intArrayOf(106, 0x110b, 0x34), // j
-            intArrayOf(107, 0x1100, 0x35), // k
-            intArrayOf(108, 0x110c, 0x36), // l
-            intArrayOf(59, 0x1107, 0x3a),
-            intArrayOf(39, 0x1110, 0x22),
-            intArrayOf(122, 0x11b7, 0x11b1), // z
-            intArrayOf(120, 0x11a8, 0x11a9), // x
-            intArrayOf(99, 0x11ae, 0x1166), // c
-            intArrayOf(118, 0x11bd, 0x1169), // v
-            intArrayOf(98, 0x116e, 0x3e), // b
-            intArrayOf(110, 0x1109, 0x30), // n
-            intArrayOf(109, 0x1112, 0x31), // m
-            intArrayOf(44, 0x2c, 0x32),
-            intArrayOf(46, 0x2e, 0x33),
-            intArrayOf(47, 0x011169, 0x3f),
-            intArrayOf(128, 0x2e, 0x2c),
-        )
-
-    @JvmField
-    val JAMO_SEBUL_3_P3_CHOJUNG =
-        arrayOf(
-            intArrayOf(49, 0x11bf, 0x21),
-            intArrayOf(50, 0x11bb, 0x40),
-            intArrayOf(51, 0x11b8, 0x23),
-            intArrayOf(52, 0x116d, 0x24),
-            intArrayOf(53, 0x1172, 0x25),
-            intArrayOf(54, 0x1163, 0x5e),
-            intArrayOf(55, 0x1168, 0x26),
-            intArrayOf(56, 0x011173, 0x2a),
-            intArrayOf(57, 0x01116e, 0x28),
-            intArrayOf(48, 0x110f, 0x29),
-            intArrayOf(113, 0x11ba, 0x11b6), // q
-            intArrayOf(119, 0x11af, 0x11b0), // w
-            intArrayOf(101, 0x1167, 0x11c0), // e
-            intArrayOf(114, 0x1165, 0x11be), // r
-            intArrayOf(116, 0x1162, 0x1164), // t
-            intArrayOf(121, 0x1105, 0x2f), // y
-            intArrayOf(117, 0x1103, 0x37), // u
-            intArrayOf(105, 0x1106, 0x38), // i
-            intArrayOf(111, 0x110e, 0x39), // o
-            intArrayOf(112, 0x1111, 0x3b), // p
-            intArrayOf(97, 0x11bc, 0x11b9), // a
-            intArrayOf(115, 0x11ab, 0x11ad), // s
-            intArrayOf(100, 0x1175, 0x11c2), // d
-            intArrayOf(102, 0x1161, 0x11c1), // f
-            intArrayOf(103, 0x1173, 0x3c), // g
-            intArrayOf(104, 0x1102, 0x27), // h
-            intArrayOf(106, 0x110b, 0x34), // j
-            intArrayOf(107, 0x1100, 0x35), // k
-            intArrayOf(108, 0x110c, 0x36), // l
-            intArrayOf(59, 0x1107, 0x3a),
-            intArrayOf(39, 0x1110, 0x22),
-            intArrayOf(122, 0x11b7, 0x11b1), // z
-            intArrayOf(120, 0x11a8, 0x11a9), // x
-            intArrayOf(99, 0x1166, 0x11ae), // c
-            intArrayOf(118, 0x1169, 0x11bd), // v
-            intArrayOf(98, 0x116e, 0x3e), // b
-            intArrayOf(110, 0x1109, 0x30), // n
-            intArrayOf(109, 0x1112, 0x31), // m
-            intArrayOf(44, 0x2c, 0x32),
-            intArrayOf(46, 0x2e, 0x33),
-            intArrayOf(47, 0x011169, 0x3f),
-            intArrayOf(128, 0x2e, 0x2c),
-        )
-	
-    // 윗글쇠 자리에 겹받침이 들어간 배열.
-    @JvmField
-    val JAMO_SEBUL_3_P3_CHOJONG_D =
-        arrayOf(
-            intArrayOf(49, 0x11bf, 0x21),
-            intArrayOf(50, 0x11bb, 0x40),
-            intArrayOf(51, 0x11b8, 0x23),
-            intArrayOf(52, 0x116d, 0x24),
-            intArrayOf(53, 0x1172, 0x25),
-            intArrayOf(54, 0x1163, 0x5e),
-            intArrayOf(55, 0x1168, 0x26),
-            intArrayOf(56, 0x011173, 0x2a),
-            intArrayOf(57, 0x01116e, 0x28),
-            intArrayOf(48, 0x110f, 0x29),
-            intArrayOf(113, 0x11ba, 0x11b6), // q
-            intArrayOf(119, 0x11af, 0x11b0), // w
-            intArrayOf(101, 0x11c0, 0x11b4), // e
-            intArrayOf(114, 0x11be, 0x11b3), // r
-            intArrayOf(116, 0x1162, 0x1164), // t
-            intArrayOf(121, 0x1105, 0x2f), // y
-            intArrayOf(117, 0x1103, 0x37), // u
-            intArrayOf(105, 0x1106, 0x38), // i
-            intArrayOf(111, 0x110e, 0x39), // o
-            intArrayOf(112, 0x1111, 0x3b), // p
-            intArrayOf(97, 0x11bc, 0x11b9), // a
-            intArrayOf(115, 0x11ab, 0x11ad), // s
-            intArrayOf(100, 0x11c2, 0x11b2), // d
-            intArrayOf(102, 0x11c1, 0x11b5), // f
-            intArrayOf(103, 0x1173, 0x3c), // g
-            intArrayOf(104, 0x1102, 0x27), // h
-            intArrayOf(106, 0x110b, 0x34), // j
-            intArrayOf(107, 0x1100, 0x35), // k
-            intArrayOf(108, 0x110c, 0x36), // l
-            intArrayOf(59, 0x1107, 0x3a),
-            intArrayOf(39, 0x1110, 0x22),
-            intArrayOf(122, 0x11b7, 0x11b1), // z
-            intArrayOf(120, 0x11a8, 0x11a9), // x
-            intArrayOf(99, 0x11ae, 0x11aa), // c
-            intArrayOf(118, 0x11bd, 0x11ac), // v
-            intArrayOf(98, 0x116e, 0x3e), // b
-            intArrayOf(110, 0x1109, 0x30), // n
-            intArrayOf(109, 0x1112, 0x31), // m
-            intArrayOf(44, 0x2c, 0x32),
-            intArrayOf(46, 0x2e, 0x33),
-            intArrayOf(47, 0x011169, 0x3f),
-            intArrayOf(128, 0x2e, 0x2c),
-        )
-
-    @JvmField
-    val JAMOSET_SEBUL_3_P3 =
-        arrayOf(
-            JAMO_SEBUL_3_P3_CHOJUNG,
-            JAMO_SEBUL_3_P3_CHOJUNG,
-            JAMO_SEBUL_3_P3_CHOJONG_D,
-            JAMO_SEBUL_3_P3_CHOJONG,
-        )
-
-    @JvmField
-    val COMB_SEBUL_3_P3 =
-        arrayOf(
-            intArrayOf(0x1162, 0x1162, 0x1164), // ㅐ + ㅐ = ㅒ
-            intArrayOf(0x1175, 0x1162, 0x1164), // ㅣ + ㅐ = ㅒ
-            intArrayOf(0x11a8, 0x11af, 0x11b0), // ㄱ + ㄹ = ㄺ
-            intArrayOf(0x11b7, 0x11af, 0x11b1), // ㅁ + ㄹ = ㄻ
-            intArrayOf(0x11b8, 0x11af, 0x11b2), // ㅂ + ㄹ = ㄼ
-            intArrayOf(0x11ba, 0x11a8, 0x11aa), // ㅅ + ㄱ = ㄳ
-            intArrayOf(0x11ba, 0x11af, 0x11b3), // ㅅ + ㄹ = ㄽ
-            intArrayOf(0x11ba, 0x11b8, 0x11b9), // ㅅ + ㅂ = ㅄ
-            intArrayOf(0x11bd, 0x11ab, 0x11ac), // ㅈ + ㄴ = ㄵ
-            intArrayOf(0x11c0, 0x11af, 0x11b4), // ㅌ + ㄹ = ㄾ
-            intArrayOf(0x11c1, 0x11af, 0x11b5), // ㅍ + ㄹ = ㄿ
-            intArrayOf(0x11c2, 0x11ab, 0x11ad), // ㅎ + ㄴ = ㄶ
-            intArrayOf(0x11c2, 0x11af, 0x11b6), // ㅎ + ㄹ = ㅀ
-            intArrayOf(0x1100, 0x1100, 0x1101),	// ㄲ
-            intArrayOf(0x1103, 0x1103, 0x1104),	// ㄸ
-            intArrayOf(0x1107, 0x1107, 0x1108),	// ㅃ
-            intArrayOf(0x1109, 0x1109, 0x110a),	// ㅆ
-            intArrayOf(0x110c, 0x110c, 0x110d),	// ㅉ
-            intArrayOf(0x1169, 0x1161, 0x116a),	// ㅘ
-            intArrayOf(0x1169, 0x1162, 0x116b),	// ㅙ
-            intArrayOf(0x1169, 0x1175, 0x116c),	// ㅚ
-            intArrayOf(0x116e, 0x1165, 0x116f),	// ㅝ
-            intArrayOf(0x116e, 0x1166, 0x1170),	// ㅞ
-            intArrayOf(0x116e, 0x1175, 0x1171),	// ㅟ
-            intArrayOf(0x1173, 0x1175, 0x1174),	// ㅢ
-            // 가상 낱자 조합
-            intArrayOf(0x011169, 0x1161, 0x116a),	// ㅘ
-            intArrayOf(0x011169, 0x1162, 0x116b),	// ㅙ
-            intArrayOf(0x011169, 0x1175, 0x116c),	// ㅚ
-            intArrayOf(0x01116e, 0x1165, 0x116f),	// ㅝ
-            intArrayOf(0x01116e, 0x1166, 0x1170),	// ㅞ
-            intArrayOf(0x01116e, 0x1175, 0x1171),	// ㅟ
-            intArrayOf(0x011173, 0x1175, 0x1174),	// ㅢ
-            intArrayOf(0x119e, 0x1175, 0x11a1),	// ㆎ
-            intArrayOf(0x119e, 0x119e, 0x11a2),	// ᆢ
-            intArrayOf(0x11a8, 0x11a8, 0x11a9),	// ㄲ
-            intArrayOf(0x11a8, 0x11ba, 0x11aa),	// ㄳ
-            intArrayOf(0x11ab, 0x11bd, 0x11ac),	// ㄵ
-            intArrayOf(0x11ab, 0x11c2, 0x11ad),	// ㄶ
-            intArrayOf(0x11af, 0x11a8, 0x11b0),	// ㄺ
-            intArrayOf(0x11af, 0x11b7, 0x11b1),	// ㄻ
-            intArrayOf(0x11af, 0x11b8, 0x11b2),	// ㄼ
-            intArrayOf(0x11af, 0x11ba, 0x11b3),	// ㄽ
-            intArrayOf(0x11af, 0x11c0, 0x11b4),	// ㄾ
-            intArrayOf(0x11af, 0x11c1, 0x11b5),	// ㄿ
-            intArrayOf(0x11af, 0x11c2, 0x11b6),	// ㅀ
-            intArrayOf(0x11b8, 0x11ba, 0x11b9),	// ㅄ
-            intArrayOf(0x11ba, 0x11ba, 0x11bb),	// ㅆ
-        )
-
-    @JvmField
-    val JAMO_SEBUL_SHIN_ORIGINAL_CHOJONG =
-        arrayOf(
-            intArrayOf(113, 0x11ba, 0x1174), // q
-            intArrayOf(119, 0x11af, 0x1163), // w
-            intArrayOf(101, 0x11b8, 0x1167), // e
-            intArrayOf(114, 0x11ae, 0x1162), // r
-            intArrayOf(116, 0x11c0, 0x1165), // t
-            intArrayOf(121, 0x1105, 0x1105), // y
-            intArrayOf(117, 0x1103, 0x1103), // u
-            intArrayOf(105, 0x1106, 0x01116e), // i
-            intArrayOf(111, 0x110e, 0x01116e), // o
-            intArrayOf(112, 0x1111, 0x011169), // p
-            intArrayOf(97, 0x11bc, 0x1164), // a
-            intArrayOf(115, 0x11ab, 0x1168), // s
-            intArrayOf(100, 0x11c2, 0x1175), // d
-            intArrayOf(102, 0x11bd, 0x1161), // f
-            intArrayOf(103, 0x11c1, 0x1173), // g
-            intArrayOf(104, 0x1102, 0x1102), // h
-            intArrayOf(106, 0x110b, 0x3b), // j
-            intArrayOf(107, 0x1100, 0x27), // k
-            intArrayOf(108, 0x110c, 0x110c), // l
-            intArrayOf(59, 0x1107, 0x3a),
-            intArrayOf(39, 0x1110, 0x22),
-            intArrayOf(122, 0x11b7, 0x1172), // z
-            intArrayOf(120, 0x11a8, 0x116d), // x
-            intArrayOf(99, 0x11be, 0x1166), // c
-            intArrayOf(118, 0x11bf, 0x1169), // v
-            intArrayOf(98, 0x11bb, 0x116e), // b
-            intArrayOf(110, 0x1109, 0x1109), // n
-            intArrayOf(109, 0x1112, 0x2f), // m
-            intArrayOf(47, 0x110f, 0x3f),
-            intArrayOf(128, 0x2e, 0x2c),
-        )
-
-    @JvmField
-    val JAMO_SEBUL_SHIN_ORIGINAL_CHOJUNG =
-        arrayOf(
-            intArrayOf(113, 0x1174, 0x11ba), // q
-            intArrayOf(119, 0x1163, 0x11af), // w
-            intArrayOf(101, 0x1167, 0x11b8), // e
-            intArrayOf(114, 0x1162, 0x11ae), // r
-            intArrayOf(116, 0x1165, 0x11c0), // t
-            intArrayOf(121, 0x1105, 0x1105), // y
-            intArrayOf(117, 0x1103, 0x1103), // u
-            intArrayOf(105, 0x01116e, 0x1106), // i
-            intArrayOf(111, 0x01116e, 0x110e), // o
-            intArrayOf(112, 0x011169, 0x1111), // p
-            intArrayOf(97, 0x1164, 0x11bc), // a
-            intArrayOf(115, 0x1168, 0x11ab), // s
-            intArrayOf(100, 0x1175, 0x11c2), // d
-            intArrayOf(102, 0x1161, 0x11bd), // f
-            intArrayOf(103, 0x1173, 0x11c1), // g
-            intArrayOf(104, 0x1102, 0x1102), // h
-            intArrayOf(106, 0x110b, 0x3b), // j
-            intArrayOf(107, 0x1100, 0x27), // k
-            intArrayOf(108, 0x110c, 0x110c), // l
-            intArrayOf(59, 0x1107, 0x3a),
-            intArrayOf(39, 0x1110, 0x22),
-            intArrayOf(122, 0x1172, 0x11b7), // z
-            intArrayOf(120, 0x116d, 0x11a8), // x
-            intArrayOf(99, 0x1166, 0x11be), // c
-            intArrayOf(118, 0x1169, 0x11bf), // v
-            intArrayOf(98, 0x116e, 0x11bb), // b
-            intArrayOf(110, 0x1109, 0x1109), // n
-            intArrayOf(109, 0x1112, 0x2f), // m
-            intArrayOf(47, 0x011169, 0x3f),
-            intArrayOf(128, 0x2e, 0x2c),
-        )
-
-    @JvmField
-    val JAMOSET_SHIN_ORIGINAL =
-        arrayOf(
-            JAMO_SEBUL_SHIN_ORIGINAL_CHOJONG,
-            JAMO_SEBUL_SHIN_ORIGINAL_CHOJUNG,
-            JAMO_SEBUL_SHIN_ORIGINAL_CHOJONG,
-            JAMO_SEBUL_SHIN_ORIGINAL_CHOJONG,
-        )
-
-    @JvmField
-    val COMB_SEBUL_SHIN_ORIGINAL =
-        arrayOf(
-            intArrayOf(0x1100, 0x1100, 0x1101),	// ㄲ
-            intArrayOf(0x1103, 0x1103, 0x1104),	// ㄸ
-            intArrayOf(0x1107, 0x1107, 0x1108),	// ㅃ
-            intArrayOf(0x1109, 0x1109, 0x110a),	// ㅆ
-            intArrayOf(0x110c, 0x110c, 0x110d),	// ㅉ
-            intArrayOf(0x1169, 0x1161, 0x116a),	// ㅘ
-            intArrayOf(0x1169, 0x1162, 0x116b),	// ㅙ
-            intArrayOf(0x1169, 0x1175, 0x116c),	// ㅚ
-            intArrayOf(0x116e, 0x1165, 0x116f),	// ㅝ
-            intArrayOf(0x116e, 0x1166, 0x1170),	// ㅞ
-            intArrayOf(0x116e, 0x1175, 0x1171),	// ㅟ
-            intArrayOf(0x1173, 0x1175, 0x1174),	// ㅢ
-            intArrayOf(0x119e, 0x1175, 0x11a1),	// ㆎ
-            intArrayOf(0x119e, 0x119e, 0x11a2),	// ᆢ
-            // 가상 낱자 조합
-            intArrayOf(0x011169, 0x1161, 0x116a),	// ㅘ
-            intArrayOf(0x011169, 0x1162, 0x116b),	// ㅙ
-            intArrayOf(0x011169, 0x1175, 0x116c),	// ㅚ
-            intArrayOf(0x01116e, 0x1165, 0x116f),	// ㅝ
-            intArrayOf(0x01116e, 0x1166, 0x1170),	// ㅞ
-            intArrayOf(0x01116e, 0x1175, 0x1171),	// ㅟ
-            intArrayOf(0x011173, 0x1175, 0x1174),	// ㅢ
-            intArrayOf(0x01119e, 0x1175, 0x11a1),	// ㆎ
-            intArrayOf(0x01119e, 0x119e, 0x11a2),	// ᆢㆎ
-            intArrayOf(0x01119e, 0x01119e, 0x11a2),	// ᆢ
-            intArrayOf(0x11a8, 0x11a8, 0x11a9),	// ㄲ
-            intArrayOf(0x11a8, 0x11ba, 0x11aa),	// ㄳ
-            intArrayOf(0x11ab, 0x11bd, 0x11ac),	// ㄵ
-            intArrayOf(0x11ab, 0x11c2, 0x11ad),	// ㄶ
-            intArrayOf(0x11af, 0x11a8, 0x11b0),	// ㄺ
-            intArrayOf(0x11af, 0x11b7, 0x11b1),	// ㄻ
-            intArrayOf(0x11af, 0x11b8, 0x11b2),	// ㄼ
-            intArrayOf(0x11af, 0x11ba, 0x11b3),	// ㄽ
-            intArrayOf(0x11af, 0x11c0, 0x11b4),	// ㄾ
-            intArrayOf(0x11af, 0x11c1, 0x11b5),	// ㄿ
-            intArrayOf(0x11af, 0x11c2, 0x11b6),	// ㅀ
-            intArrayOf(0x11b8, 0x11ba, 0x11b9),	// ㅄ
-            intArrayOf(0x11ba, 0x11ba, 0x11bb),	// ㅆ
-        )
-
-    @JvmField
-    val JAMO_SEBUL_SHIN_EDIT_CHOJONG =
-        arrayOf(
-            intArrayOf(113, 0x11ba, 0x1164), // q
-            intArrayOf(119, 0x11af, 0x1163), // w
-            intArrayOf(101, 0x11b8, 0x1167), // e
-            intArrayOf(114, 0x11c0, 0x1162), // r
-            intArrayOf(116, 0x11c1, 0x1165), // t
-            intArrayOf(121, 0x1105, 0x201c), // y
-            intArrayOf(117, 0x1103, 0x201d), // u
-            intArrayOf(105, 0x1106, 0x1174), // i
-            intArrayOf(111, 0x110e, 0x01116e), // o
-            intArrayOf(112, 0x1111, 0x011169), // p
-            intArrayOf(97, 0x11bc, 0x1172), // a
-            intArrayOf(115, 0x11ab, 0x1168), // s
-            intArrayOf(100, 0x11ae, 0x1175), // d
-            intArrayOf(102, 0x11bb, 0x1161), // f
-            intArrayOf(103, 0x11bd, 0x1173), // g
-            intArrayOf(104, 0x1102, 0x2018), // h
-            intArrayOf(106, 0x110b, 0x2019), // j
-            intArrayOf(107, 0x1100, 0x3b), // k
-            intArrayOf(108, 0x110c, 0x27), // l
-            intArrayOf(59, 0x1107, 0x3a),
-            intArrayOf(39, 0x1110, 0x22),
-            intArrayOf(122, 0x11b7, 0x203b), // z
-            intArrayOf(120, 0x11a8, 0x116d), // x
-            intArrayOf(99, 0x11be, 0x1166), // c
-            intArrayOf(118, 0x11c2, 0x1169), // v
-            intArrayOf(98, 0x11bf, 0x116e), // b
-            intArrayOf(110, 0x1109, 0x00b7), // n
-            intArrayOf(109, 0x1112, 0x2f), // m
-            intArrayOf(47, 0x110f, 0x3f),
-            intArrayOf(128, 0x2e, 0x2c),
-        )
-
-    @JvmField
-    val JAMO_SEBUL_SHIN_EDIT_CHOJUNG =
-        arrayOf(
-            intArrayOf(113, 0x1164, 0x11ba), // q
-            intArrayOf(119, 0x1163, 0x11af), // w
-            intArrayOf(101, 0x1167, 0x11b8), // e
-            intArrayOf(114, 0x1162, 0x11c0), // r
-            intArrayOf(116, 0x1165, 0x11c1), // t
-            intArrayOf(121, 0x1105, 0x201c), // y
-            intArrayOf(117, 0x1103, 0x201d), // u
-            intArrayOf(105, 0x1174, 0x1106), // i
-            intArrayOf(111, 0x01116e, 0x110e), // o
-            intArrayOf(112, 0x011169, 0x1111), // p
-            intArrayOf(97, 0x1172, 0x11bc), // a
-            intArrayOf(115, 0x1168, 0x11ab), // s
-            intArrayOf(100, 0x1175, 0x11ae), // d
-            intArrayOf(102, 0x1161, 0x11bb), // f
-            intArrayOf(103, 0x1173, 0x11bd), // g
-            intArrayOf(104, 0x1102, 0x2018), // h
-            intArrayOf(106, 0x110b, 0x2019), // j
-            intArrayOf(107, 0x1100, 0x3b), // k
-            intArrayOf(108, 0x110c, 0x27), // l
-            intArrayOf(59, 0x1107, 0x3a),
-            intArrayOf(39, 0x1110, 0x22),
-            intArrayOf(122, 0x203b, 0x11b7), // z
-            intArrayOf(120, 0x116d, 0x11a8), // x
-            intArrayOf(99, 0x1166, 0x11be), // c
-            intArrayOf(118, 0x1169, 0x11c2), // v
-            intArrayOf(98, 0x116e, 0x11bf), // b
-            intArrayOf(110, 0x1109, 0x00b7), // n
-            intArrayOf(109, 0x1112, 0x2f), // m
-            intArrayOf(47, 0x011169, 0x3f),
-            intArrayOf(128, 0x2e, 0x2c),
-        )
-
-    @JvmField
-    val JAMOSET_SHIN_EDIT =
-        arrayOf(
-            JAMO_SEBUL_SHIN_EDIT_CHOJONG,
-            JAMO_SEBUL_SHIN_EDIT_CHOJUNG,
-            JAMO_SEBUL_SHIN_EDIT_CHOJONG,
-            JAMO_SEBUL_SHIN_EDIT_CHOJONG,
-        )
-
-    @JvmField
-    val JAMO_SEBUL_SHIN_M_CHOJONG =
-        arrayOf(
-            intArrayOf(113, 0x11ba, 0x1164), // q
-            intArrayOf(119, 0x11af, 0x1163), // w
-            intArrayOf(101, 0x11b8, 0x1167), // e
-            intArrayOf(114, 0x11bd, 0x1162), // r
-            intArrayOf(116, 0x11bf, 0x1165), // t
-            intArrayOf(121, 0x1105, 0x201c), // y
-            intArrayOf(117, 0x1103, 0x201d), // u
-            intArrayOf(105, 0x1106, 0x203b), // i
-            intArrayOf(111, 0x110e, 0x01116e), // o
-            intArrayOf(112, 0x1111, 0x011169), // p
-            intArrayOf(97, 0x11bc, 0x1172), // a
-            intArrayOf(115, 0x11ab, 0x1174), // s
-            intArrayOf(100, 0x11bb, 0x1175), // d
-            intArrayOf(102, 0x11c0, 0x1161), // f
-            intArrayOf(103, 0x11ae, 0x1173), // g
-            intArrayOf(104, 0x1102, 0x300a), // h
-            intArrayOf(106, 0x110b, 0x300b), // j
-            intArrayOf(107, 0x1100, 0xb7), // k
-            intArrayOf(108, 0x110c, 0x3b), // l
-            intArrayOf(59, 0x1107, 0x3a), // ;
-            intArrayOf(39, 0x1110, 0x2f), // '
-            intArrayOf(122, 0x11b7, 0x1168), // z
-            intArrayOf(120, 0x11a8, 0x116d), // x
-            intArrayOf(99, 0x11c2, 0x1166), // c
-            intArrayOf(118, 0x11c1, 0x1169), // v
-            intArrayOf(98, 0x11be, 0x116e), // b
-            intArrayOf(110, 0x1109, 0x27), // n
-            intArrayOf(109, 0x1112, 0x22), // m
-            intArrayOf(44, 0x2c, 0x3c), // ,
-            intArrayOf(46, 0x2e, 0x3e), // .
-            intArrayOf(47, 0x110f, 0x3f), // /
-            intArrayOf(128, 0x2e, 0x2c),
-        )
-
-    @JvmField
-    val JAMO_SEBUL_SHIN_M_CHOJUNG =
-        arrayOf(
-            intArrayOf(113, 0x1164, 0x11ba), // q
-            intArrayOf(119, 0x1163, 0x11af), // w
-            intArrayOf(101, 0x1167, 0x11b8), // e
-            intArrayOf(114, 0x1162, 0x11bd), // r
-            intArrayOf(116, 0x1165, 0x11bf), // t
-            intArrayOf(121, 0x1105, 0x201c), // y
-            intArrayOf(117, 0x1103, 0x201d), // u
-            intArrayOf(105, 0x1106, 0x203b), // i
-            intArrayOf(111, 0x110e, 0x110e), // o
-            intArrayOf(112, 0x1111, 0x1111), // p
-            intArrayOf(97, 0x1172, 0x11bc), // a
-            intArrayOf(115, 0x1174, 0x11ab), // s
-            intArrayOf(100, 0x1175, 0x11bb), // d
-            intArrayOf(102, 0x1161, 0x11c0), // f
-            intArrayOf(103, 0x1173, 0x11ae), // g
-            intArrayOf(104, 0x1102, 0x300a), // h
-            intArrayOf(106, 0x110b, 0x300b), // j
-            intArrayOf(107, 0x1100, 0xb7), // k
-            intArrayOf(108, 0x110c, 0x3b), // l
-            intArrayOf(59, 0x1107, 0x3a), // ;
-            intArrayOf(39, 0x1110, 0x2f), // '
-            intArrayOf(122, 0x1168, 0x11b7), // z
-            intArrayOf(120, 0x116d, 0x11a8), // x
-            intArrayOf(99, 0x1166, 0x11c2), // c
-            intArrayOf(118, 0x1169, 0x11c1), // v
-            intArrayOf(98, 0x116e, 0x11be), // b
-            intArrayOf(110, 0x1109, 0x27), // n
-            intArrayOf(109, 0x1112, 0x22), // m
-            intArrayOf(44, 0x2c, 0x3c), // ,
-            intArrayOf(46, 0x2e, 0x3e), // .
-            intArrayOf(47, 0x110f, 0x110f), // /
-            intArrayOf(128, 0x2e, 0x2c),
-        )
-
-    @JvmField
-    val JAMO_SEBUL_SHIN_M_CHODJUNG =
-        arrayOf(
-            intArrayOf(113, 0x1164, 0x11ba), // q
-            intArrayOf(119, 0x1163, 0x11af), // w
-            intArrayOf(101, 0x1167, 0x11b8), // e
-            intArrayOf(114, 0x1162, 0x11bd), // r
-            intArrayOf(116, 0x1165, 0x11bf), // t
-            intArrayOf(121, 0x1105, 0x201c), // y
-            intArrayOf(117, 0x1103, 0x201d), // u
-            intArrayOf(105, 0x1106, 0x203b), // i
-            intArrayOf(111, 0x01116e, 0x110e), // o
-            intArrayOf(112, 0x011169, 0x1111), // p
-            intArrayOf(97, 0x1172, 0x11bc), // a
-            intArrayOf(115, 0x1174, 0x11ab), // s
-            intArrayOf(100, 0x1175, 0x11bb), // d
-            intArrayOf(102, 0x1161, 0x11c0), // f
-            intArrayOf(103, 0x1173, 0x11ae), // g
-            intArrayOf(104, 0x1102, 0x300a), // h
-            intArrayOf(106, 0x110b, 0x300b), // j
-            intArrayOf(107, 0x1100, 0xb7), // k
-            intArrayOf(108, 0x110c, 0x3b), // l
-            intArrayOf(59, 0x1107, 0x3a), // ;
-            intArrayOf(39, 0x1110, 0x2f), // '
-            intArrayOf(122, 0x1168, 0x11b7), // z
-            intArrayOf(120, 0x116d, 0x11a8), // x
-            intArrayOf(99, 0x1166, 0x11c2), // c
-            intArrayOf(118, 0x1169, 0x11c1), // v
-            intArrayOf(98, 0x116e, 0x11be), // b
-            intArrayOf(110, 0x1109, 0x27), // n
-            intArrayOf(109, 0x1112, 0x22), // m
-            intArrayOf(44, 0x2c, 0x3c), // ,
-            intArrayOf(46, 0x2e, 0x3e), // .
-            intArrayOf(47, 0x011169, 0x110f), // /
-            intArrayOf(128, 0x2e, 0x2c),
-        )
-
-    @JvmField
-    val JAMOSET_SHIN_M =
-        arrayOf(
-            JAMO_SEBUL_SHIN_M_CHOJUNG,
-            JAMO_SEBUL_SHIN_M_CHODJUNG,
-            JAMO_SEBUL_SHIN_M_CHOJONG,
-            JAMO_SEBUL_SHIN_M_CHOJONG,
-        )
-
-    @JvmField
-    val JAMO_SEBUL_SHIN_P2_CHOJONG =
-        arrayOf(
-            intArrayOf(113, 0x11ba, 0x1164), // q
-            intArrayOf(119, 0x11af, 0x1163), // w
-            intArrayOf(101, 0x11b8, 0x1162), // e
-            intArrayOf(114, 0x11c0, 0x1165), // r
-            intArrayOf(116, 0x11bf, 0x1167), // t
-            intArrayOf(121, 0x1105, 0x302f), // y
-            intArrayOf(117, 0x1103, 0x302e), // u
-            intArrayOf(105, 0x1106, 0x011173), // i
-            intArrayOf(111, 0x110e, 0x01116e), // o
-            intArrayOf(112, 0x1111, 0x01119e), // p
-            intArrayOf(97, 0x11bc, 0x1172), // a
-            intArrayOf(115, 0x11ab, 0x1168), // s
-            intArrayOf(100, 0x11c2, 0x1175), // d
-            intArrayOf(102, 0x11c1, 0x1161), // f
-            intArrayOf(103, 0x11ae, 0x1173), // g
-            intArrayOf(104, 0x1102, 0x25a1), // h
-            intArrayOf(106, 0x110b, 0x27), // j
-            intArrayOf(107, 0x1100, 0xb7), // k
-            intArrayOf(108, 0x110c, 0x3b), // l
-            intArrayOf(59, 0x1107, 0x3a),
-            intArrayOf(39, 0x1110, 0x2f),
-            intArrayOf(122, 0x11b7, 0x119e), // z
-            intArrayOf(120, 0x11bb, 0x116d), // x
-            intArrayOf(99, 0x11a8, 0x1166), // c
-            intArrayOf(118, 0x11bd, 0x1169), // v
-            intArrayOf(98, 0x11be, 0x116e), // b
-            intArrayOf(110, 0x1109, 0x2015), // n
-            intArrayOf(109, 0x1112, 0x22), // m
-            intArrayOf(44, 0x2c, 0x3c),
-            intArrayOf(46, 0x2e, 0x3e),
-            intArrayOf(47, 0x110f, 0x3f),
-            intArrayOf(128, 0x2e, 0x2c),
-        )
-
-    @JvmField
-    val JAMO_SEBUL_SHIN_P2_CHOJUNG =
-        arrayOf(
-            intArrayOf(113, 0x1164, 0x11ba), // q
-            intArrayOf(119, 0x1163, 0x11af), // w
-            intArrayOf(101, 0x1162, 0x11b8), // e
-            intArrayOf(114, 0x1165, 0x11c0), // r
-            intArrayOf(116, 0x1167, 0x11bf), // t
-            intArrayOf(121, 0x1105, 0x302f), // y
-            intArrayOf(117, 0x1103, 0x302e), // u
-            intArrayOf(105, 0x011173, 0x1106), // i
-            intArrayOf(111, 0x01116e, 0x110e), // o
-            intArrayOf(112, 0x01119e, 0x1111), // p
-            intArrayOf(97, 0x1172, 0x11bc), // a
-            intArrayOf(115, 0x1168, 0x11ab), // s
-            intArrayOf(100, 0x1175, 0x11c2), // d
-            intArrayOf(102, 0x1161, 0x11c1), // f
-            intArrayOf(103, 0x1173, 0x11ae), // g
-            intArrayOf(104, 0x1102, 0x25a1), // h
-            intArrayOf(106, 0x110b, 0x27), // j
-            intArrayOf(107, 0x1100, 0xb7), // k
-            intArrayOf(108, 0x110c, 0x3b), // l
-            intArrayOf(59, 0x1107, 0x3a),
-            intArrayOf(39, 0x1110, 0x2f),
-            intArrayOf(122, 0x119e, 0x11b7), // z
-            intArrayOf(120, 0x116d, 0x11bb), // x
-            intArrayOf(99, 0x1166, 0x11a8), // c
-            intArrayOf(118, 0x1169, 0x11bd), // v
-            intArrayOf(98, 0x116e, 0x11be), // b
-            intArrayOf(110, 0x1109, 0x2015), // n
-            intArrayOf(109, 0x1112, 0x22), // m
-            intArrayOf(44, 0x2c, 0x3c),
-            intArrayOf(46, 0x2e, 0x3e),
-            intArrayOf(47, 0x011169, 0x110f),
-            intArrayOf(128, 0x2e, 0x2c),
-        )
-
-    @JvmField
-    val JAMOSET_SHIN_P2 =
-        arrayOf(
-            JAMO_SEBUL_SHIN_P2_CHOJONG,
-            JAMO_SEBUL_SHIN_P2_CHOJUNG,
-            JAMO_SEBUL_SHIN_P2_CHOJONG,
-            JAMO_SEBUL_SHIN_P2_CHOJONG,
-        )
-}
