--- app/src/main/java/io/github/lens0021/teogeul/TeogeulKOKR.kt
@@ -1,820 +1,0 @@
-package io.github.lens0021.teogeul
-
-import android.content.Context
-import android.content.Intent
-import android.content.res.Configuration
-import android.os.Handler
-import android.os.IBinder
-import android.os.Looper
-import android.text.method.MetaKeyKeyListener
-import android.view.KeyEvent
-import android.view.View
-import android.view.inputmethod.EditorInfo
-import android.view.inputmethod.InputMethodManager
-import io.github.lens0021.teogeul.KOKR.EngineMode
-import io.github.lens0021.teogeul.KOKR.HangulEngine
-import io.github.lens0021.teogeul.KOKR.HangulEngine.FinishComposingEvent
-import io.github.lens0021.teogeul.KOKR.HangulEngine.HangulEngineEvent
-import io.github.lens0021.teogeul.KOKR.HangulEngine.HangulEngineListener
-import io.github.lens0021.teogeul.KOKR.HangulEngine.SetComposingEvent
-import io.github.lens0021.teogeul.KOKR.ListLangKeyActionDialogActivity
-import io.github.lens0021.teogeul.data.SettingsDefaults
-import io.github.lens0021.teogeul.data.SettingsRepository
-import io.github.lens0021.teogeul.data.SettingsValues
-import io.github.lens0021.teogeul.data.settingsDataStore
-import io.github.lens0021.teogeul.event.CommitComposingTextEvent
-import io.github.lens0021.teogeul.event.InputKeyEvent
-import io.github.lens0021.teogeul.event.InputTimeoutEvent
-import io.github.lens0021.teogeul.event.KeyUpEvent
-import io.github.lens0021.teogeul.event.TeogeulEvent
-import io.github.lens0021.teogeul.event.TeogeulEventFlow
-import io.github.lens0021.teogeul.ui.TeogeulSettingsActivity
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.Job
-import kotlinx.coroutines.SupervisorJob
-import kotlinx.coroutines.cancel
-import kotlinx.coroutines.flow.collect
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.runBlocking
-
-class TeogeulKOKR : Teogeul, HangulEngineListener {
-    companion object {
-        @JvmField
-        val SHIFT_CONVERT =
-            arrayOf(
-                intArrayOf(0x60, 0x7e),
-                intArrayOf(0x31, 0x21),
-                intArrayOf(0x32, 0x40),
-                intArrayOf(0x33, 0x23),
-                intArrayOf(0x34, 0x24),
-                intArrayOf(0x35, 0x25),
-                intArrayOf(0x36, 0x5e),
-                intArrayOf(0x37, 0x26),
-                intArrayOf(0x38, 0x2a),
-                intArrayOf(0x39, 0x28),
-                intArrayOf(0x30, 0x29),
-                intArrayOf(0x2d, 0x5f),
-                intArrayOf(0x3d, 0x2b),
-                intArrayOf(0x5b, 0x7b),
-                intArrayOf(0x5d, 0x7d),
-                intArrayOf(0x5c, 0x7c),
-                intArrayOf(0x3b, 0x3a),
-                intArrayOf(0x27, 0x22),
-                intArrayOf(0x2c, 0x3c),
-                intArrayOf(0x2e, 0x3e),
-                intArrayOf(0x2f, 0x3f),
-            )
-
-        private val QWERTY_TO_DVORAK: MutableMap<Int, Int> =
-            HashMap<Int, Int>().apply {
-                put(KeyEvent.KEYCODE_Q, KeyEvent.KEYCODE_APOSTROPHE)
-                put(KeyEvent.KEYCODE_W, KeyEvent.KEYCODE_COMMA)
-                put(KeyEvent.KEYCODE_E, KeyEvent.KEYCODE_PERIOD)
-                put(KeyEvent.KEYCODE_R, KeyEvent.KEYCODE_P)
-                put(KeyEvent.KEYCODE_T, KeyEvent.KEYCODE_Y)
-                put(KeyEvent.KEYCODE_Y, KeyEvent.KEYCODE_F)
-                put(KeyEvent.KEYCODE_U, KeyEvent.KEYCODE_G)
-                put(KeyEvent.KEYCODE_I, KeyEvent.KEYCODE_C)
-                put(KeyEvent.KEYCODE_O, KeyEvent.KEYCODE_R)
-                put(KeyEvent.KEYCODE_P, KeyEvent.KEYCODE_L)
-                put(KeyEvent.KEYCODE_LEFT_BRACKET, KeyEvent.KEYCODE_SLASH)
-                put(KeyEvent.KEYCODE_RIGHT_BRACKET, KeyEvent.KEYCODE_EQUALS)
-
-                put(KeyEvent.KEYCODE_A, KeyEvent.KEYCODE_A)
-                put(KeyEvent.KEYCODE_S, KeyEvent.KEYCODE_O)
-                put(KeyEvent.KEYCODE_D, KeyEvent.KEYCODE_E)
-                put(KeyEvent.KEYCODE_F, KeyEvent.KEYCODE_U)
-                put(KeyEvent.KEYCODE_G, KeyEvent.KEYCODE_I)
-                put(KeyEvent.KEYCODE_H, KeyEvent.KEYCODE_D)
-                put(KeyEvent.KEYCODE_J, KeyEvent.KEYCODE_H)
-                put(KeyEvent.KEYCODE_K, KeyEvent.KEYCODE_T)
-                put(KeyEvent.KEYCODE_L, KeyEvent.KEYCODE_N)
-                put(KeyEvent.KEYCODE_SEMICOLON, KeyEvent.KEYCODE_S)
-                put(KeyEvent.KEYCODE_APOSTROPHE, KeyEvent.KEYCODE_MINUS)
-
-                put(KeyEvent.KEYCODE_Z, KeyEvent.KEYCODE_SEMICOLON)
-                put(KeyEvent.KEYCODE_X, KeyEvent.KEYCODE_Q)
-                put(KeyEvent.KEYCODE_C, KeyEvent.KEYCODE_J)
-                put(KeyEvent.KEYCODE_V, KeyEvent.KEYCODE_K)
-                put(KeyEvent.KEYCODE_B, KeyEvent.KEYCODE_X)
-                put(KeyEvent.KEYCODE_N, KeyEvent.KEYCODE_B)
-                put(KeyEvent.KEYCODE_M, KeyEvent.KEYCODE_M)
-                put(KeyEvent.KEYCODE_COMMA, KeyEvent.KEYCODE_W)
-                put(KeyEvent.KEYCODE_PERIOD, KeyEvent.KEYCODE_V)
-                put(KeyEvent.KEYCODE_SLASH, KeyEvent.KEYCODE_Z)
-
-                put(KeyEvent.KEYCODE_MINUS, KeyEvent.KEYCODE_LEFT_BRACKET)
-                put(KeyEvent.KEYCODE_EQUALS, KeyEvent.KEYCODE_RIGHT_BRACKET)
-            }
-
-        private val QWERTY_TO_COLEMAK: MutableMap<Int, Int> =
-            HashMap<Int, Int>().apply {
-                put(KeyEvent.KEYCODE_E, KeyEvent.KEYCODE_F)
-                put(KeyEvent.KEYCODE_R, KeyEvent.KEYCODE_P)
-                put(KeyEvent.KEYCODE_T, KeyEvent.KEYCODE_G)
-                put(KeyEvent.KEYCODE_Y, KeyEvent.KEYCODE_J)
-                put(KeyEvent.KEYCODE_U, KeyEvent.KEYCODE_L)
-                put(KeyEvent.KEYCODE_I, KeyEvent.KEYCODE_U)
-                put(KeyEvent.KEYCODE_O, KeyEvent.KEYCODE_Y)
-                put(KeyEvent.KEYCODE_P, KeyEvent.KEYCODE_SEMICOLON)
-
-                put(KeyEvent.KEYCODE_S, KeyEvent.KEYCODE_R)
-                put(KeyEvent.KEYCODE_D, KeyEvent.KEYCODE_S)
-                put(KeyEvent.KEYCODE_F, KeyEvent.KEYCODE_T)
-                put(KeyEvent.KEYCODE_G, KeyEvent.KEYCODE_D)
-                put(KeyEvent.KEYCODE_J, KeyEvent.KEYCODE_N)
-                put(KeyEvent.KEYCODE_K, KeyEvent.KEYCODE_E)
-                put(KeyEvent.KEYCODE_L, KeyEvent.KEYCODE_I)
-                put(KeyEvent.KEYCODE_SEMICOLON, KeyEvent.KEYCODE_O)
-
-                put(KeyEvent.KEYCODE_N, KeyEvent.KEYCODE_K)
-            }
-
-        const val LANGKEY_LIST_ACTIONS = "list_actions"
-        const val LANGKEY_SWITCH_KOR_ENG = "switch_kor_eng"
-        const val LANGKEY_SWITCH_NEXT_METHOD = "switch_next_method"
-        const val LANGKEY_LIST_METHODS = "list_methods"
-        const val LANGKEY_OPEN_SETTINGS = "open_settings"
-
-        private var mSelf: TeogeulKOKR? = null
-
-        @JvmStatic
-        fun getInstance(): TeogeulKOKR? {
-            return mSelf
-        }
-    }
-
-    var mHangulEngine: HangulEngine
-    var mQwertyEngine: HangulEngine
-
-    var mHardShift: Int = 0
-    var mHardAlt: Int = 0
-    var mCapsLock: Boolean = false
-    var mShiftOnCapsLock: Boolean = false
-
-    var mShiftPressing: Boolean = false
-    var mAltPressing: Boolean = false
-
-    lateinit var mCurrentEngineMode: EngineMode
-    var mCurrentLanguage: Int = EngineMode.LANG_KO
-
-    private val mShiftKeyToggle = intArrayOf(0, MetaKeyKeyListener.META_SHIFT_ON, MetaKeyKeyListener.META_CAP_LOCKED)
-    private val mAltKeyToggle = intArrayOf(0, MetaKeyKeyListener.META_ALT_ON, MetaKeyKeyListener.META_ALT_LOCKED)
-
-    var mEnableTimeout: Boolean = false
-
-    var mMoachigi: Boolean = false
-    var mHardwareMoachigi: Boolean = false
-    var mFullMoachigi: Boolean = true
-    var mMoachigiDelay: Int = 0
-
-    var mStandardJamo: Boolean = false
-
-    var mAlphabetLayout: String = "keyboard_alphabet_qwerty"
-
-    var mSelectionMode: Boolean = false
-    var mSelectionStart: Int = 0
-    var mSelectionEnd: Int = 0
-
-    var mInput: Boolean = false
-
-    var mTimeOutHandler: Handler? = null
-
-    var mHardLangKey: KeyStroke? = null
-
-    private val mAltLayout: Array<IntArray> = emptyArray()
-    private val eventScope = CoroutineScope(SupervisorJob() + Dispatchers.Unconfined)
-    private var eventJob: Job? = null
-    private val settingsRepository by lazy { SettingsRepository(applicationContext.settingsDataStore) }
-
-    constructor() : super() {
-        mSelf = this
-
-        mQwertyEngine = HangulEngine()
-        mQwertyEngine.listener = this
-        mHangulEngine = mQwertyEngine
-
-        mAutoHideMode = false
-    }
-
-    constructor(context: Context) : this() {
-        attachBaseContext(context)
-    }
-
-    override fun onCreate() {
-        super.onCreate()
-        eventJob =
-            eventScope.launch {
-                TeogeulEventFlow.events.collect { event ->
-                    when (event) {
-                        is InputKeyEvent -> handleInputKey(event)
-                        is KeyUpEvent -> handleKeyUp(event)
-                        is InputTimeoutEvent -> handleInputTimeout(event)
-                        is CommitComposingTextEvent -> handleCommitComposingText(event)
-                    }
-                }
-            }
-    }
-
-    override fun onCreateInputView(): View? {
-        return super.onCreateInputView()
-    }
-
-    override fun onStartInputView(
-        attribute: EditorInfo,
-        restarting: Boolean,
-    ) {
-        resetCharComposition()
-        super.onStartInputView(attribute, restarting)
-        applyPreferences(resetLanguage = true)
-    }
-
-    override fun onStartInput(
-        attribute: EditorInfo,
-        restarting: Boolean,
-    ) {
-        super.onStartInput(attribute, restarting)
-        applyPreferences(resetLanguage = true)
-    }
-
-    override fun onConfigurationChanged(newConfig: Configuration) {
-        super.onConfigurationChanged(newConfig)
-    }
-
-    override fun onFinishInput() {
-        resetCharComposition()
-        super.onFinishInput()
-    }
-
-    // onViewClicked is deprecated and no longer called since API 28
-    // Keeping empty implementation for compatibility
-    override fun onViewClicked(focusChanged: Boolean) {
-        // No-op: This callback is deprecated and not invoked on modern Android versions
-    }
-
-    override fun onEvent(event: HangulEngineEvent) {
-        if (event is FinishComposingEvent) {
-            mInputConnection?.finishComposingText()
-        }
-        if (event is SetComposingEvent) {
-            mInputConnection?.setComposingText(event.composing, 1)
-        }
-    }
-
-    private fun handleKeyUp(event: KeyUpEvent) {
-        val key = event.keyEvent.keyCode
-        if (!mShiftPressing) {
-            if (key == KeyEvent.KEYCODE_SHIFT_LEFT || key == KeyEvent.KEYCODE_SHIFT_RIGHT) {
-                mHardShift = 0
-                mShiftPressing = true
-                if (mShiftOnCapsLock) {
-                    mHardShift = 2
-                    mShiftPressing = true
-                    mShiftOnCapsLock = false
-                }
-                updateMetaKeyStateDisplay()
-            }
-        }
-        // Alt key is not handled by IME - let system handle it
-    }
-
-    private fun handleInputTimeout(
-        @Suppress("UNUSED_PARAMETER") event: InputTimeoutEvent,
-    ) {
-        if (mEnableTimeout) {
-            resetCharComposition()
-        }
-    }
-
-    private fun handleInputKey(event: InputKeyEvent) {
-        val keyEvent = event.keyEvent
-        when (keyEvent.keyCode) {
-            // Alt key is not handled by IME - let system handle it for shortcuts
-
-            KeyEvent.KEYCODE_SHIFT_LEFT, KeyEvent.KEYCODE_SHIFT_RIGHT -> {
-                if (keyEvent.repeatCount == 0) {
-                    if (++mHardShift > 2) {
-                        mHardShift = 0
-                    }
-                }
-                mShiftPressing = true
-                if (mCapsLock) {
-                    mHardShift = 0
-                    mShiftPressing = false
-                    mShiftOnCapsLock = true
-                }
-                updateMetaKeyStateDisplay()
-                return
-            }
-
-            KeyEvent.KEYCODE_CAPS_LOCK -> {
-                mCapsLock = !mCapsLock
-                if (mCapsLock) {
-                    mHardShift = 2
-                    mShiftPressing = true
-                } else {
-                    mHardShift = 0
-                    mShiftPressing = false
-                }
-                updateMetaKeyStateDisplay()
-                return
-            }
-        }
-        if (keyEvent.metaState and KeyEvent.META_CAPS_LOCK_ON != 0) {
-            if (!mShiftOnCapsLock) {
-                mCapsLock = true
-                mHardShift = 2
-                mShiftPressing = true
-                updateMetaKeyStateDisplay()
-            }
-        } else if (mCapsLock) {
-            mCapsLock = false
-            mHardShift = 0
-            mShiftPressing = false
-            updateMetaKeyStateDisplay()
-        }
-        val ret = processKeyEvent(keyEvent)
-        event.isCancelled = ret
-        shinShift()
-    }
-
-    private fun handleCommitComposingText(
-        @Suppress("UNUSED_PARAMETER") event: CommitComposingTextEvent,
-    ) {
-        currentInputConnection.finishComposingText()
-    }
-
-    override fun onEvent(ev: TeogeulEvent): Boolean {
-        return when (ev) {
-            is InputKeyEvent -> processKeyEvent(ev.keyEvent)
-            else -> false
-        }
-    }
-
-    private fun inputChar(code: Char) {
-        var shift = mHardShift
-        var mutableCode = code
-        var isDirect = false
-
-        if (mutableCode.code == 128) {
-            mutableCode = if (shift > 0) 0x2c.toChar() else 0x2e.toChar()
-            shift = 0
-            isDirect = true
-        }
-
-        val originalCode = mutableCode
-        for (item in SHIFT_CONVERT) {
-            if (mutableCode.code == item[1]) {
-                mutableCode = item[0].toChar()
-                shift = 1
-            }
-        }
-
-        if (mDirectInputMode) {
-            mutableCode = originalCode
-            resetCharComposition()
-            directInput(mutableCode, shift > 0)
-            return
-        } else if (isDirect) {
-            resetCharComposition()
-            mInputConnection?.commitText(String(charArrayOf(originalCode)), 1)
-            resetCharComposition()
-            return
-        }
-
-        val inputCode = mHangulEngine.inputCode(mutableCode.lowercaseChar().code, shift)
-        if (inputCode != -1) {
-            if (mHangulEngine.inputJamo(inputCode) == 0) {
-                mInputConnection?.commitText(String(charArrayOf(inputCode.toChar())), 1)
-                resetCharComposition()
-            }
-        } else {
-            resetCharComposition()
-            if (shift > 0) {
-                mutableCode = originalCode.uppercaseChar()
-                for (item in SHIFT_CONVERT) {
-                    if (mutableCode.code == item[0]) {
-                        mutableCode = item[1].toChar()
-                    }
-                }
-            }
-            mInputConnection?.commitText(String(charArrayOf(mutableCode)), 1)
-            resetCharComposition()
-        }
-    }
-
-    private fun directInput(
-        code: Char,
-        shift: Boolean,
-    ) {
-        var mutableCode = code
-        if (shift) {
-            mutableCode = mutableCode.uppercaseChar()
-            for (item in SHIFT_CONVERT) {
-                if (mutableCode.code == item[0]) {
-                    mutableCode = item[1].toChar()
-                    break
-                }
-            }
-        }
-        mInputConnection?.commitText(mutableCode.toString(), 1)
-    }
-
-    fun onLangKey(action: String) {
-        val imm = getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager?
-        val token: IBinder = requireNotNull(window?.window).attributes.token
-
-        when (action) {
-            LANGKEY_LIST_ACTIONS -> {
-                val intent = Intent(this, ListLangKeyActionDialogActivity::class.java)
-                intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-                startActivity(intent)
-            }
-
-            LANGKEY_SWITCH_KOR_ENG -> {
-                toggleLanguage()
-            }
-
-            LANGKEY_SWITCH_NEXT_METHOD -> {
-                // Note: switchToLastInputMethod and switchToNextInputMethod are deprecated in API 28,
-                // but no replacement API is available. These methods still work on modern Android.
-                if (mInput) {
-                    mInput = false
-                    imm?.switchToLastInputMethod(token)
-                } else {
-                    imm?.switchToNextInputMethod(token, false)
-                }
-            }
-
-            LANGKEY_LIST_METHODS -> {
-                imm?.showInputMethodPicker()
-            }
-
-            LANGKEY_OPEN_SETTINGS -> {
-                val intent = Intent(this, TeogeulSettingsActivity::class.java)
-                intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-                startActivity(intent)
-            }
-        }
-    }
-
-    private fun convertQwertyToLayout(ev: KeyEvent): KeyEvent? {
-        val conversionMap =
-            when (mAlphabetLayout) {
-                "keyboard_alphabet_dvorak" -> QWERTY_TO_DVORAK
-                "keyboard_alphabet_colemak" -> QWERTY_TO_COLEMAK
-                else -> return null // QWERTY or unknown - no conversion needed
-            }
-
-        val originalKeyCode = ev.keyCode
-        val convertedKeyCode = conversionMap[originalKeyCode] ?: return null
-
-        return KeyEvent(
-            ev.downTime,
-            ev.eventTime,
-            ev.action,
-            convertedKeyCode,
-            ev.repeatCount,
-            ev.metaState,
-            ev.deviceId,
-            ev.scanCode,
-            ev.flags,
-            ev.source,
-        )
-    }
-
-    private fun processKeyEvent(ev: KeyEvent): Boolean {
-        val inputConnection = mInputConnection ?: return false
-        val key = ev.keyCode
-
-        if (ev.isShiftPressed) {
-            when (key) {
-                KeyEvent.KEYCODE_DPAD_UP,
-                KeyEvent.KEYCODE_DPAD_DOWN,
-                KeyEvent.KEYCODE_DPAD_LEFT,
-                KeyEvent.KEYCODE_DPAD_RIGHT,
-                -> {
-                    if (!mSelectionMode) {
-                        val beforeAll = inputConnection.getTextBeforeCursor(Int.MAX_VALUE, 0)
-                        mSelectionEnd = beforeAll?.length ?: 0
-                        mSelectionStart = mSelectionEnd
-                        mSelectionMode = true
-                    } else {
-                        if (key == KeyEvent.KEYCODE_DPAD_LEFT) mSelectionEnd--
-                        if (key == KeyEvent.KEYCODE_DPAD_RIGHT) mSelectionEnd++
-                        if (key == KeyEvent.KEYCODE_DPAD_UP) {
-                            var i = 1
-                            var text: CharSequence = ""
-                            var end: Boolean
-                            while (true) {
-                                val before = inputConnection.getTextBeforeCursor(i, 0) ?: ""
-                                end = before == text
-                                text = before
-                                if (end || (text.isNotEmpty() && text[0] == '\n')) {
-                                    break
-                                }
-                                i++
-                            }
-                            if (end) {
-                                val beforeAll = inputConnection.getTextBeforeCursor(Int.MAX_VALUE, 0)
-                                mSelectionEnd -= beforeAll?.length ?: 0
-                            } else {
-                                mSelectionEnd -= i
-                            }
-                        }
-                        if (key == KeyEvent.KEYCODE_DPAD_DOWN) {
-                            var i = 1
-                            var text: CharSequence = ""
-                            var end: Boolean
-                            while (true) {
-                                val after = inputConnection.getTextAfterCursor(i, 0) ?: ""
-                                end = after == text
-                                text = after
-                                if (end || (text.isNotEmpty() && text[text.length - 1] == '\n')) {
-                                    break
-                                }
-                                i++
-                            }
-                            if (end) {
-                                val afterAll = inputConnection.getTextAfterCursor(Char.MAX_VALUE.code, 0)
-                                mSelectionEnd += afterAll?.length ?: 0
-                            } else {
-                                mSelectionEnd += i
-                            }
-                        }
-                        var start = mSelectionStart
-                        var end = mSelectionEnd
-                        if (mSelectionStart > mSelectionEnd) {
-                            start = mSelectionEnd
-                            end = mSelectionStart
-                        }
-                        inputConnection.setSelection(start, end)
-                        mHardShift = 0
-                        updateMetaKeyStateDisplay()
-                        updateMetaKeyStateDisplay()
-                    }
-                    return true
-                }
-
-                else -> {
-                    mSelectionMode = false
-                }
-            }
-        } else {
-            mSelectionMode = false
-        }
-
-        // Ctrl key handling (available since API 11, always true for minSdk 26)
-        if (ev.isCtrlPressed) {
-            val convertedEvent = convertQwertyToLayout(ev)
-            if (convertedEvent != null) {
-                inputConnection.sendKeyEvent(convertedEvent)
-                return true
-            }
-            return false
-        }
-
-        // Alt key is not handled by IME - let system handle it for shortcuts
-        if (ev.isAltPressed) {
-            return false
-        }
-
-        if (key >= KeyEvent.KEYCODE_NUMPAD_0 && key <= KeyEvent.KEYCODE_NUMPAD_RIGHT_PAREN) {
-            resetCharComposition()
-            return false
-        }
-
-        // Language switch key is not handled by IME - let system handle it
-        if (key == KeyEvent.KEYCODE_LANGUAGE_SWITCH) {
-            resetCharComposition()
-            return false
-        }
-
-        // Handle custom language key combination (e.g., user-defined shortcut for toggling language)
-        val hardLangKey = mHardLangKey
-        if (hardLangKey != null && key == hardLangKey.keyCode) {
-            if ((mHardShift == 1) == hardLangKey.shift &&
-                ev.isAltPressed == hardLangKey.alt &&
-                ev.isCtrlPressed == hardLangKey.control &&
-                ev.isMetaPressed == hardLangKey.win
-            ) {
-                resetCharComposition()
-                toggleLanguage()
-                mHardShift = 0
-                mShiftPressing = false
-                updateMetaKeyStateDisplay()
-                return true
-            }
-        }
-
-        if (ev.isPrintingKey) {
-            var processedEvent: KeyEvent = ev
-            val convertedEvent = convertQwertyToLayout(ev)
-            if (convertedEvent != null) {
-                processedEvent = convertedEvent
-            }
-
-            // Don't apply Alt meta state for character input - Alt is only used for shortcuts/language switching
-            val code = processedEvent.getUnicodeChar(mShiftKeyToggle[mHardShift])
-            inputChar(code.toChar())
-            mInput = true
-
-            if (mHardShift == 1) {
-                mShiftPressing = false
-            }
-            if (!ev.isShiftPressed) {
-                if (mHardShift == 1) {
-                    mHardShift = 0
-                }
-            }
-            if (!ev.isShiftPressed) {
-                updateMetaKeyStateDisplay()
-            }
-
-            return true
-        } else if (key == KeyEvent.KEYCODE_SPACE) {
-            resetCharComposition()
-            inputConnection.commitText(" ", 1)
-            return true
-        } else if (key == KeyEvent.KEYCODE_DEL) {
-            if (!mHangulEngine.backspace()) {
-                resetCharComposition()
-                inputConnection.deleteSurroundingText(1, 0)
-            }
-            if (mHangulEngine.composing == "") {
-                resetCharComposition()
-            }
-            return true
-        } else if (key == KeyEvent.KEYCODE_ENTER) {
-            resetCharComposition()
-            mHardShift = 0
-            updateMetaKeyStateDisplay()
-            val editorInfo = currentInputEditorInfo
-            return when (editorInfo.imeOptions and EditorInfo.IME_MASK_ACTION) {
-                EditorInfo.IME_ACTION_SEARCH, EditorInfo.IME_ACTION_GO -> {
-                    sendDefaultEditorAction(true)
-                    true
-                }
-
-                else -> false
-            }
-        } else {
-            resetCharComposition()
-        }
-
-        return false
-    }
-
-    private fun shinShift() {
-        updateMetaKeyStateDisplay()
-    }
-
-    fun updateMetaKeyStateDisplay() {
-    }
-
-    private fun applyPreferences(resetLanguage: Boolean = false) {
-        val snapshot = runBlocking { settingsRepository.snapshot() }
-        mHardwareMoachigi = snapshot.hardwareUseMoachigi
-        mFullMoachigi = snapshot.hardwareFullMoachigi
-        mMoachigiDelay = snapshot.hardwareFullMoachigiDelay
-        mStandardJamo = snapshot.systemUseStandardJamo
-        if (resetLanguage) {
-            applyStartLanguage(snapshot.systemStartHangulMode)
-        }
-        mHardLangKey = KeyStroke.parse(snapshot.systemHardwareLangKeyStroke)
-
-        // Use subtype layout when explicitly set; otherwise use app settings.
-        val imm = getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager?
-        val currentSubtype = imm?.currentInputMethodSubtype
-        val subtypeLayout =
-            currentSubtype?.extraValue?.let { extraValue ->
-                Regex("KeyboardLayoutSet=(\\w+)").find(extraValue)?.groupValues?.get(1)
-            }
-        mAlphabetLayout =
-            if (subtypeLayout != null && subtypeLayout != SettingsDefaults.HARDWARE_ALPHABET_LAYOUT) {
-                subtypeLayout
-            } else {
-                snapshot.hardwareAlphabetLayout
-            }
-
-        val modeKey =
-            if (mCurrentLanguage == EngineMode.LANG_EN) {
-                mAlphabetLayout
-            } else {
-                snapshot.hardwareHangulLayout
-            }
-        applyEngineMode(EngineMode.get(modeKey))
-    }
-
-    private fun applyStartLanguage(mode: String) {
-        when (mode) {
-            SettingsValues.START_HANGUL_MODE_START_HANGUL ->
-                mCurrentLanguage = EngineMode.LANG_KO
-            SettingsValues.START_HANGUL_MODE_START_ENGLISH ->
-                mCurrentLanguage = EngineMode.LANG_EN
-            "new_input",
-            "always",
-            -> mCurrentLanguage = EngineMode.LANG_KO
-        }
-    }
-
-    private fun applyEngineMode(mode: EngineMode) {
-        mCurrentEngineMode = mode
-        if (mode == EngineMode.DIRECT) {
-            mDirectInputMode = true
-            mEnableTimeout = false
-            mFullMoachigi = false
-            mHangulEngine = mQwertyEngine
-            mHangulEngine.jamoTable = null
-            mHangulEngine.setCombinationTable(null)
-            return
-        }
-
-        val prop = mode.properties
-        mDirectInputMode = prop.direct
-        mEnableTimeout = prop.timeout
-        mFullMoachigi = prop.fullMoachigi
-        mHangulEngine = mQwertyEngine
-        if (mode.jamoset != null) {
-            mHangulEngine.jamoSet = mode.jamoset
-        } else {
-            mHangulEngine.jamoTable = mode.layout
-        }
-        mHangulEngine.setCombinationTable(mode.combination)
-        mHangulEngine.firstMidEnd = mStandardJamo
-        mHangulEngine.moachigi = mHardwareMoachigi
-        mHangulEngine.fullMoachigi = mFullMoachigi
-        if (mFullMoachigi) {
-            mEnableTimeout = true
-        }
-    }
-
-    private fun resetCharComposition() {
-        mHangulEngine.resetComposition()
-    }
-
-    private fun toggleLanguage() {
-        mCurrentLanguage =
-            if (mCurrentLanguage == EngineMode.LANG_EN) {
-                EngineMode.LANG_KO
-            } else {
-                EngineMode.LANG_EN
-            }
-        applyPreferences()
-    }
-
-    fun resetHardShift(force: Boolean) {
-        if (mHardShift == 2 && !force) {
-            return
-        }
-        mHardShift = 0
-        mShiftPressing = false
-    }
-
-    override fun hideWindow() {
-        super.hideWindow()
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        eventJob?.cancel()
-        eventScope.cancel()
-    }
-
-    override fun onKeyDown(
-        keyCode: Int,
-        event: KeyEvent,
-    ): Boolean {
-        if (mTimeOutHandler == null) {
-            mTimeOutHandler = Handler(Looper.getMainLooper())
-            mTimeOutHandler?.postDelayed({
-                TeogeulEventFlow.emitBlocking(InputTimeoutEvent())
-                mTimeOutHandler = null
-            }, mMoachigiDelay.toLong())
-        }
-        return super.onKeyDown(keyCode, event)
-    }
-
-    override fun onEvaluateFullscreenMode(): Boolean {
-        return false
-    }
-
-    override fun onEvaluateInputViewShown(): Boolean {
-        super.onEvaluateInputViewShown()
-        return true
-    }
-
-    override fun requestHideSelf(flag: Int) {
-        super.requestHideSelf(flag)
-        hideWindow()
-    }
-
-    fun getHangulEngine(): HangulEngine {
-        return mHangulEngine
-    }
-
-    fun getAltSymbols(): Array<IntArray> {
-        return mAltLayout
-    }
-}
