+++ app/src/main/java/io/github/lens0021/teogeul/compose/LetterConverter.kt
@@ -1,6 +1,6 @@
-package io.github.lens0021.teogeul
+package io.github.lens0021.teogeul.compose
 
-import io.github.lens0021.teogeul.data.SettingsSnapshot
+import io.github.lens0021.teogeul.config.SettingsSnapshot
 
 interface LetterConverter {
     fun convert(text: ComposingText): Boolean
