+++ app/src/main/java/io/github/lens0021/teogeul/compose/StrSegment.kt
@@ -1,4 +1,4 @@
-package io.github.lens0021.teogeul
+package io.github.lens0021.teogeul.compose
 
 class StrSegment {
     var string: String?
