--- app/src/main/java/io/github/lens0021/teogeul/event/CommitComposingTextEvent.kt
@@ -1,3 +1,0 @@
-package io.github.lens0021.teogeul.event
-
-class CommitComposingTextEvent : TeogeulEvent()
