--- app/src/main/java/io/github/lens0021/teogeul/event/InputKeyEvent.kt
@@ -1,5 +1,0 @@
-package io.github.lens0021.teogeul.event
-
-import android.view.KeyEvent
-
-class InputKeyEvent(val keyEvent: KeyEvent) : TeogeulEvent()
