--- app/src/main/java/io/github/lens0021/teogeul/event/TeogeulEvent.kt
@@ -1,5 +1,0 @@
-package io.github.lens0021.teogeul.event
-
-abstract class TeogeulEvent {
-    var isCancelled: Boolean = false
-}
