+++ app/src/main/java/io/github/lens0021/teogeul/korean/EngineMode.kt
@@ -1,11 +1,6 @@
-package io.github.lens0021.teogeul.KOKR
+package io.github.lens0021.teogeul.korean
 
-import io.github.lens0021.teogeul.KOKR.layout.LayoutAlphabet
-import io.github.lens0021.teogeul.KOKR.layout.LayoutDev
-import io.github.lens0021.teogeul.KOKR.layout.LayoutDubul
-import io.github.lens0021.teogeul.KOKR.layout.LayoutGongSebul
-import io.github.lens0021.teogeul.KOKR.layout.LayoutMoachigiSebul
-import io.github.lens0021.teogeul.KOKR.layout.LayoutShinSebul
+import io.github.lens0021.teogeul.layout.LayoutData
 
 private const val LANG_EN_CODE = 0
 private const val LANG_KO_CODE = 3
@@ -22,151 +17,151 @@ enum class EngineMode(
 
     SEBUL_390(
         Properties(),
-        LayoutGongSebul.JAMO_SEBUL_390,
+        LayoutData.JAMO_SEBUL_390,
         null,
-        LayoutGongSebul.COMB_SEBULSIK,
+        LayoutData.COMB_SEBULSIK,
         null,
         "keyboard_sebul_390",
     ),
     SEBUL_391(
         Properties(),
-        LayoutGongSebul.JAMO_SEBUL_391,
+        LayoutData.JAMO_SEBUL_391,
         null,
-        LayoutGongSebul.COMB_SEBULSIK,
+        LayoutData.COMB_SEBULSIK,
         null,
         "keyboard_sebul_391",
     ),
     SEBUL_DANMOEUM(
         Properties(),
-        LayoutGongSebul.JAMO_SEBUL_DANMOEUM,
+        LayoutData.JAMO_SEBUL_DANMOEUM,
         null,
-        LayoutGongSebul.COMB_SEBUL_DANMOEUM,
+        LayoutData.COMB_SEBUL_DANMOEUM,
         null,
         "keyboard_sebul_danmoeum",
     ),
     DUBULSIK(
         Properties(),
-        LayoutDubul.JAMO_DUBUL_STANDARD,
+        LayoutData.JAMO_DUBUL_STANDARD,
         null,
-        LayoutDubul.COMB_DUBUL_STANDARD,
+        LayoutData.COMB_DUBUL_STANDARD,
         null,
         "keyboard_dubul_standard",
     ),
     DUBULSIK_NK(
         Properties(),
-        LayoutDubul.JAMO_DUBUL_NK,
+        LayoutData.JAMO_DUBUL_NK,
         null,
-        LayoutDubul.COMB_DUBUL_STANDARD,
+        LayoutData.COMB_DUBUL_STANDARD,
         null,
         "keyboard_dubul_nk",
     ),
 
     SEBUL_SUN_2014(
         Properties(),
-        LayoutGongSebul.JAMO_SEBUL_SUN_2014,
+        LayoutData.JAMO_SEBUL_SUN_2014,
         null,
-        LayoutGongSebul.COMB_SEBUL_SUN_2014,
+        LayoutData.COMB_SEBUL_SUN_2014,
         null,
         "keyboard_sebul_sun_2014",
     ),
     SEBUL_3_2015M(
         Properties(),
         null,
-        LayoutShinSebul.JAMOSET_SEBUL_3_2015M,
-        LayoutShinSebul.COMB_SEBUL_3_2015,
+        LayoutData.JAMOSET_SEBUL_3_2015M,
+        LayoutData.COMB_SEBUL_3_2015,
         null,
         "keyboard_sebul_3_2015m",
     ),
     SEBUL_3_2015(
         Properties(),
         null,
-        LayoutShinSebul.JAMOSET_SEBUL_3_2015,
-        LayoutShinSebul.COMB_SEBUL_3_2015,
+        LayoutData.JAMOSET_SEBUL_3_2015,
+        LayoutData.COMB_SEBUL_3_2015,
         null,
         "keyboard_sebul_3_2015",
     ),
     SEBUL_3_P3(
         Properties(),
         null,
-        LayoutShinSebul.JAMOSET_SEBUL_3_P3,
-        LayoutShinSebul.COMB_SEBUL_3_P3,
+        LayoutData.JAMOSET_SEBUL_3_P3,
+        LayoutData.COMB_SEBUL_3_P3,
         null,
         "keyboard_sebul_3_p3",
     ),
     SEBUL_SHIN_ORIGINAL(
         Properties(),
         null,
-        LayoutShinSebul.JAMOSET_SHIN_ORIGINAL,
-        LayoutShinSebul.COMB_SEBUL_SHIN_ORIGINAL,
+        LayoutData.JAMOSET_SHIN_ORIGINAL,
+        LayoutData.COMB_SEBUL_SHIN_ORIGINAL,
         null,
         "keyboard_sebul_shin_original",
     ),
     SEBUL_SHIN_EDIT(
         Properties(),
         null,
-        LayoutShinSebul.JAMOSET_SHIN_EDIT,
-        LayoutShinSebul.COMB_SEBUL_SHIN_ORIGINAL,
+        LayoutData.JAMOSET_SHIN_EDIT,
+        LayoutData.COMB_SEBUL_SHIN_ORIGINAL,
         null,
         "keyboard_sebul_shin_edit",
     ),
     SEBUL_SHIN_M(
         Properties(),
         null,
-        LayoutShinSebul.JAMOSET_SHIN_M,
-        LayoutShinSebul.COMB_SEBUL_SHIN_ORIGINAL,
+        LayoutData.JAMOSET_SHIN_M,
+        LayoutData.COMB_SEBUL_SHIN_ORIGINAL,
         null,
         "keyboard_sebul_shin_m",
     ),
     SEBUL_SHIN_P2(
         Properties(),
         null,
-        LayoutShinSebul.JAMOSET_SHIN_P2,
-        LayoutGongSebul.COMB_FULL,
+        LayoutData.JAMOSET_SHIN_P2,
+        LayoutData.COMB_FULL,
         null,
         "keyboard_sebul_shin_p2",
     ),
 
     SEBUL_AHNMATAE(
         Properties(true, false, false, false),
-        LayoutMoachigiSebul.JAMO_SEBUL_AHNMATAE,
+        LayoutData.JAMO_SEBUL_AHNMATAE,
         null,
-        LayoutMoachigiSebul.COMB_SEBUL_AHNMATAE,
+        LayoutData.COMB_SEBUL_AHNMATAE,
         null,
         "keyboard_sebul_ahnmatae",
     ),
     SEBUL_SEMOE_2016(
         Properties(true, false, false, false),
-        LayoutMoachigiSebul.JAMO_SEBUL_SEMOE_2016,
+        LayoutData.JAMO_SEBUL_SEMOE_2016,
         null,
-        LayoutMoachigiSebul.COMB_SEBUL_SEMOE,
+        LayoutData.COMB_SEBUL_SEMOE,
         null,
         "keyboard_sebul_semoe_2016",
     ),
     SEBUL_SEMOE(
         Properties(true, false, false, false),
-        LayoutMoachigiSebul.JAMO_SEBUL_SEMOE_2016,
+        LayoutData.JAMO_SEBUL_SEMOE_2016,
         null,
-        LayoutMoachigiSebul.COMB_SEBUL_SEMOE,
+        LayoutData.COMB_SEBUL_SEMOE,
         null,
         "keyboard_sebul_semoe",
     ),
 
-    NEBUL_1969(Properties(), LayoutDev.JAMO_NEBUL_1969, null, LayoutDev.COMB_NEBUL_1969, null, "keyboard_nebul_1969"),
+    NEBUL_1969(Properties(), LayoutData.JAMO_NEBUL_1969, null, LayoutData.COMB_NEBUL_1969, null, "keyboard_nebul_1969"),
 
-    DUBULSIK_YET(Properties(), LayoutDubul.JAMO_DUBUL_YET, null, LayoutGongSebul.COMB_FULL, null, "keyboard_dubul_yet"),
+    DUBULSIK_YET(Properties(), LayoutData.JAMO_DUBUL_YET, null, LayoutData.COMB_FULL, null, "keyboard_dubul_yet"),
     SEBUL_393Y(
         Properties(),
-        LayoutGongSebul.JAMO_SEBUL_393Y,
+        LayoutData.JAMO_SEBUL_393Y,
         null,
-        LayoutGongSebul.COMB_FULL,
+        LayoutData.COMB_FULL,
         null,
         "keyboard_sebul_393y",
     ),
     SEBUL_3_2015Y(
         Properties(),
-        LayoutShinSebul.JAMO_SEBUL_3_2015Y,
+        LayoutData.JAMO_SEBUL_3_2015Y,
         null,
-        LayoutGongSebul.COMB_FULL,
+        LayoutData.COMB_FULL,
         null,
         "keyboard_sebul_3_2015y",
     ),
@@ -181,7 +176,7 @@ enum class EngineMode(
     ),
     ENGLISH_DVORAK(
         Properties(LANG_EN_CODE, false, false, false, false, false, false),
-        LayoutAlphabet.CONVERT_ENGLISH_DVORAK,
+        LayoutData.CONVERT_ENGLISH_DVORAK,
         null,
         null,
         null,
@@ -189,7 +184,7 @@ enum class EngineMode(
     ),
     ENGLISH_COLEMAK(
         Properties(LANG_EN_CODE, false, false, false, false, false, false),
-        LayoutAlphabet.CONVERT_ENGLISH_COLEMAK,
+        LayoutData.CONVERT_ENGLISH_COLEMAK,
         null,
         null,
         null,
