+++ app/src/main/java/io/github/lens0021/teogeul/korean/LangKeyActionDialog.kt
@@ -1,13 +1,13 @@
-package io.github.lens0021.teogeul.KOKR
+package io.github.lens0021.teogeul.korean
 
 import android.app.Activity
 import android.app.AlertDialog
 import android.content.DialogInterface
 import android.os.Bundle
 import io.github.lens0021.teogeul.R
-import io.github.lens0021.teogeul.TeogeulKOKR
+import io.github.lens0021.teogeul.IMEService
 
-class ListLangKeyActionDialogActivity : Activity() {
+class LangKeyActionDialog : Activity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         val labels =
@@ -19,16 +19,16 @@ class ListLangKeyActionDialogActivity : Activity() {
             )
         val actions =
             arrayOf(
-                TeogeulKOKR.LANGKEY_SWITCH_KOR_ENG,
-                TeogeulKOKR.LANGKEY_SWITCH_NEXT_METHOD,
-                TeogeulKOKR.LANGKEY_LIST_METHODS,
-                TeogeulKOKR.LANGKEY_OPEN_SETTINGS,
+                IMEService.LANGKEY_SWITCH_KOR_ENG,
+                IMEService.LANGKEY_SWITCH_NEXT_METHOD,
+                IMEService.LANGKEY_LIST_METHODS,
+                IMEService.LANGKEY_OPEN_SETTINGS,
             )
         val builder = AlertDialog.Builder(this, android.R.style.Theme_Material_Light_Dialog)
         builder
             .setTitle(R.string.preference_system_list_actions_title)
             .setItems(labels) { dialog: DialogInterface, which: Int ->
-                TeogeulKOKR.getInstance()?.onLangKey(actions[which])
+                IMEService.getInstance()?.onLangKey(actions[which])
                 dialog.dismiss()
                 finish()
             }
