+++ app/src/main/java/io/github/lens0021/teogeul/korean/LayoutRegistry.kt
@@ -1,0 +1,58 @@
+package io.github.lens0021.teogeul.korean
+
+import io.github.lens0021.teogeul.layout.LayoutData
+
+object LayoutRegistry {
+    fun getLayout(name: String): Array<IntArray>? {
+        return when (name) {
+            "JAMO_SEBUL_3_2015M_CHOJONG" -> LayoutData.JAMO_SEBUL_3_2015M_CHOJONG
+            "JAMO_SEBUL_3_2015M_CHOJUNG" -> LayoutData.JAMO_SEBUL_3_2015M_CHOJUNG
+            "JAMO_SEBUL_3_2015M_CHOJONG_D" -> LayoutData.JAMO_SEBUL_3_2015M_CHOJONG_D
+            "JAMO_SEBUL_3_2015_CHOJONG" -> LayoutData.JAMO_SEBUL_3_2015_CHOJONG
+            "JAMO_SEBUL_3_2015_CHOJUNG" -> LayoutData.JAMO_SEBUL_3_2015_CHOJUNG
+            "JAMO_SEBUL_3_2015_CHOJONG_D" -> LayoutData.JAMO_SEBUL_3_2015_CHOJONG_D
+            "JAMO_SEBUL_3_2015Y" -> LayoutData.JAMO_SEBUL_3_2015Y
+            "COMB_SEBUL_3_2015" -> LayoutData.COMB_SEBUL_3_2015
+            "JAMO_SEBUL_3_P3_CHOJONG" -> LayoutData.JAMO_SEBUL_3_P3_CHOJONG
+            "JAMO_SEBUL_3_P3_CHOJUNG" -> LayoutData.JAMO_SEBUL_3_P3_CHOJUNG
+            "JAMO_SEBUL_3_P3_CHOJONG_D" -> LayoutData.JAMO_SEBUL_3_P3_CHOJONG_D
+            "COMB_SEBUL_3_P3" -> LayoutData.COMB_SEBUL_3_P3
+            "JAMO_SEBUL_SHIN_ORIGINAL_CHOJONG" -> LayoutData.JAMO_SEBUL_SHIN_ORIGINAL_CHOJONG
+            "JAMO_SEBUL_SHIN_ORIGINAL_CHOJUNG" -> LayoutData.JAMO_SEBUL_SHIN_ORIGINAL_CHOJUNG
+            "COMB_SEBUL_SHIN_ORIGINAL" -> LayoutData.COMB_SEBUL_SHIN_ORIGINAL
+            "JAMO_SEBUL_SHIN_EDIT_CHOJONG" -> LayoutData.JAMO_SEBUL_SHIN_EDIT_CHOJONG
+            "JAMO_SEBUL_SHIN_EDIT_CHOJUNG" -> LayoutData.JAMO_SEBUL_SHIN_EDIT_CHOJUNG
+            "JAMO_SEBUL_SHIN_M_CHOJONG" -> LayoutData.JAMO_SEBUL_SHIN_M_CHOJONG
+            "JAMO_SEBUL_SHIN_M_CHOJUNG" -> LayoutData.JAMO_SEBUL_SHIN_M_CHOJUNG
+            "JAMO_SEBUL_SHIN_M_CHODJUNG" -> LayoutData.JAMO_SEBUL_SHIN_M_CHODJUNG
+            "JAMO_SEBUL_SHIN_P2_CHOJONG" -> LayoutData.JAMO_SEBUL_SHIN_P2_CHOJONG
+            "JAMO_SEBUL_SHIN_P2_CHOJUNG" -> LayoutData.JAMO_SEBUL_SHIN_P2_CHOJUNG
+            "JAMO_SEBUL_DANMOEUM" -> LayoutData.JAMO_SEBUL_DANMOEUM
+            "COMB_SEBUL_DANMOEUM" -> LayoutData.COMB_SEBUL_DANMOEUM
+            "JAMO_SEBUL_391" -> LayoutData.JAMO_SEBUL_391
+            "JAMO_SEBUL_390" -> LayoutData.JAMO_SEBUL_390
+            "JAMO_SEBUL_393Y" -> LayoutData.JAMO_SEBUL_393Y
+            "JAMO_SEBUL_SUN_2014" -> LayoutData.JAMO_SEBUL_SUN_2014
+            "COMB_SEBUL_SUN_2014" -> LayoutData.COMB_SEBUL_SUN_2014
+            "COMB_SEBULSIK" -> LayoutData.COMB_SEBULSIK
+            "COMB_FULL" -> LayoutData.COMB_FULL
+            "JAMO_SEBUL_AHNMATAE" -> LayoutData.JAMO_SEBUL_AHNMATAE
+            "COMB_SEBUL_AHNMATAE" -> LayoutData.COMB_SEBUL_AHNMATAE
+            "JAMO_SEBUL_SEMOE_2016" -> LayoutData.JAMO_SEBUL_SEMOE_2016
+            "JAMO_SEBUL_SEMOE" -> LayoutData.JAMO_SEBUL_SEMOE
+            "COMB_SEBUL_SEMOE" -> LayoutData.COMB_SEBUL_SEMOE
+            "JAMO_DUBUL_STANDARD" -> LayoutData.JAMO_DUBUL_STANDARD
+            "COMB_DUBUL_STANDARD" -> LayoutData.COMB_DUBUL_STANDARD
+            "JAMO_DUBUL_NK" -> LayoutData.JAMO_DUBUL_NK
+            "JAMO_DUBUL_YET" -> LayoutData.JAMO_DUBUL_YET
+            "JAMO_DUBUL_DANMOEUM_GOOGLE" -> LayoutData.JAMO_DUBUL_DANMOEUM_GOOGLE
+            "COMB_DUBUL_DANMOEUM_GOOGLE" -> LayoutData.COMB_DUBUL_DANMOEUM_GOOGLE
+            "JAMO_NEBUL_1969" -> LayoutData.JAMO_NEBUL_1969
+            "COMB_NEBUL_1969" -> LayoutData.COMB_NEBUL_1969
+            "VIRTUAL_NEBUL_1969" -> LayoutData.VIRTUAL_NEBUL_1969
+            "CONVERT_ENGLISH_DVORAK" -> LayoutData.CONVERT_ENGLISH_DVORAK
+            "CONVERT_ENGLISH_COLEMAK" -> LayoutData.CONVERT_ENGLISH_COLEMAK
+            else -> null
+        }
+    }
+}
