+++ app/src/main/java/io/github/lens0021/teogeul/model/Key.kt
@@ -1,4 +1,4 @@
-package io.github.lens0021.teogeul
+package io.github.lens0021.teogeul.model
 
 open class Key {
     var isPressed: Boolean = false
