+++ app/src/main/java/io/github/lens0021/teogeul/model/KeyStroke.kt
@@ -1,4 +1,4 @@
-package io.github.lens0021.teogeul
+package io.github.lens0021.teogeul.model
 
 data class KeyStroke(
     val control: Boolean,
