+++ app/src/main/java/io/github/lens0021/teogeul/model/ModifierKey.kt
@@ -1,4 +1,6 @@
-package io.github.lens0021.teogeul
+package io.github.lens0021.teogeul.model
+
+import io.github.lens0021.teogeul.ModifierState
 
 open class ModifierKey : Key() {
     var state: ModifierState = ModifierState.Disabled
