+++ app/src/main/java/io/github/lens0021/teogeul/ui/Preferences.kt
@@ -20,7 +20,7 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.viewinterop.AndroidView
 import androidx.compose.ui.window.Dialog
-import io.github.lens0021.teogeul.KeyStroke
+import io.github.lens0021.teogeul.model.KeyStroke
 import io.github.lens0021.teogeul.R
 
 // Utility
